% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_addscale.R
\name{bb_addscale}
\alias{bb_addscale}
\title{adds color scale for heatmap plots}
\usage{
bb_addscale(color_vector, min_val, max_val, border = FALSE, height,
  width, x, y, units = "inches", orientation = "vertical",
  fontsize = 8, fontcolor = "grey", fontface = "plain",
  just = c("left", "top"), ...)
}
\arguments{
\item{color_vector}{sequential list of colors used to represent interaction frequency; can be returned from bb_plothic function}

\item{min_val}{label for minimum value, i.e first value of zrange}

\item{max_val}{label for maximum value, i.e. last value of zrange}

\item{border}{option to add border around scale}

\item{height}{scale height}

\item{width}{scale width}

\item{x}{scale x-coordinate}

\item{y}{scale y-coordinate}

\item{units}{units of height, width, and x and y coordinates}

\item{orientation}{"vertical" or "horizontal" orientation}

\item{fontsize}{fontsize for text}

\item{fontcolor}{fontcolor for text}

\item{fontface}{fontface for text}

\item{just}{justification of scale viewport}
}
\description{
adds color scale for heatmap plots
}
\author{
Nicole Kramer
}
