% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_addText.R
\name{bb_addText}
\alias{bb_addText}
\title{wrapper to draw a textGrob based on bb_makePage coordinates and units}
\usage{
bb_addText(
  label,
  x,
  y,
  just = "center",
  rotation = 0,
  fontcolor = "black",
  transparency = 1,
  fontsize = 12,
  cex = 1,
  fontfamily = "",
  fontface = "plain",
  default.units = "inches"
)
}
\arguments{
\item{label}{character or expression}

\item{x}{A numeric vector or unit object specifying x-location}

\item{y}{A numeric vector or unit object specifying y-location}

\item{just}{justification of text relative to its (x, y) location}

\item{rotation}{angle to rotate text}

\item{fontcolor}{fontcolor}

\item{transparency}{degree of text transparency}

\item{fontsize}{the size of text (in points)}

\item{cex}{multiplier applied to fontsize}

\item{fontfamily}{the font family}

\item{fontface}{the fontface (bold, italic, ...)}

\item{default.units}{A string indicating the default units to use if x or y are only given as numeric vectors}
}
\description{
wrapper to draw a textGrob based on bb_makePage coordinates and units
}
