% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_annoBedpeAnchors.R
\name{bb_annoBedpeAnchors}
\alias{bb_annoBedpeAnchors}
\title{annotates vertical boxes at anchors of bedpe elements}
\usage{
bb_annoBedpeAnchors(
  plot,
  x,
  y,
  height,
  params = NULL,
  fillcolor = "lightgrey",
  linecolor = NA,
  alpha = 0.4,
  just = c("left", "top"),
  default.units = "inches"
)
}
\arguments{
\item{plot}{bedpe plot to highlight bedpe anchors}

\item{x}{A numeric or unit object specifying x-location.}

\item{y}{A numeric or unit object specifying y-location.}

\item{height}{A numeric or unit object specifying height.}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{fillcolor}{character value specifying fillcolor of boxes}

\item{linecolor}{character value specifying linecolor of boxes}

\item{alpha}{numeric value specifying transparency}

\item{just}{justification of text relative to its (x, y) location}

\item{default.units}{A string indicating the default units to use if x or y are only given as numerics.}
}
\description{
annotates vertical boxes at anchors of bedpe elements
}
