% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_annoLoops.R
\name{bb_annoLoops}
\alias{bb_annoLoops}
\title{annotates loops in a Hi-C plot}
\usage{
bb_annoLoops(
  hic,
  loops,
  params = NULL,
  half = "inherit",
  shift = 4,
  type = "box",
  ...
)
}
\arguments{
\item{hic}{hic plot to annotate}

\item{loops}{bedpe file or dataframe in bedpe file format with loop positions}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{half}{which half of hic plots to annotate; default is "inherit", which will inherit whatever is plotted; other options are "both", "top", or "bottom"}

\item{shift}{number of pixels on either end of loop in box/circle; number of pixels for length of arrow}

\item{type}{type of annotation; options are "box", "circle", or "arrow"}
}
\description{
annotates loops in a Hi-C plot
}
