% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_annoPixels.R
\name{bb_annoPixels}
\alias{bb_annoPixels}
\title{Annotate pixels in a Hi-C plot}
\usage{
bb_annoPixels(
  plot,
  data,
  type = "box",
  half = "inherit",
  shift = 4,
  params = NULL,
  ...
)
}
\arguments{
\item{plot}{Hi-C plot object from \code{bb_plotHicSquare} or \code{bb_plotHicTriangle} on which to annotate pixels.}

\item{data}{A string specifying the BEDPE file path or a dataframe in BEDPE format specifying pixel positions.}

\item{type}{Character value specifying type of annotation. Default value is \code{type = "box"}. Options are:
\itemize{
\item{\code{"box"}: }{Boxes are drawn around each pixel.}
\item{\code{"circle"}: }{Circles are drawn around each pixel.}
\item{\code{"arrow"}: }{Arrows are drawn pointing to each pixel.}
}}

\item{half}{Character value specifying which half of hic plots to annotate. Triangle Hi-C plots will always default to the entirety of the triangular plot. Default value is \code{half = "inherit"}. Options are:
\itemize{
\item{\code{"inherit"}: }{Pixels will be annotated on the \code{half} inherited by the input Hi-C plot.}
\item{\code{"both"}: }{Pixels will be annotated on both halves of the diagonal of a square Hi-C plot.}
\item{\code{"top"}: }{Pixels will be annotated on the upper diagonal half of a square Hi-C plot.}
\item{\code{"bottom"}: }{Pixels will be annotated ont the bottom diagonal half of a square Hi-C plot.}
}}

\item{shift}{Numeric specifying the number of pixels on either end of main pixel in a box or circle. Numeric specifying number of pixels for the length of an arrow.}

\item{params}{An optional \link[BentoBox]{bb_params} object containing relevant function parameters.}

\item{...}{Additional grid graphical parameters. See \link[grid]{gpar}.}
}
\value{
Returns a \code{bb_pixel} object containing relevant genomic region, placement, and \link[grid]{grob} information.
}
\description{
Annotate pixels in a Hi-C plot
}
\examples{
## Load Hi-C data and BEDPE data
data("bb_hicData")
data("bb_bedpeData")

## Create BentoBox page
bb_pageCreate(width = 4, height = 4, default.units = "inches")

## Plot and place a square Hi-C plot
hicPlot <- bb_plotHicSquare(data = bb_hicData, resolution = 10000, zrange = c(0, 70),
                            chrom = "chr21", chromstart = 28000000, chromend = 30300000,
                            x = 0.5, y = 0.5, width = 3, height = 3, just = c("left", "top"),
                            default.units = "inches")

## Annotate loops of Hi-C plot
bb_annoPixels(plot = hicPlot, data = bb_bedpeData, type = "box", half = "both")

## Annotate genome label
bb_annoGenomeLabel(plot = hicPlot, x = 0.5, y = 3.6, scale = "Mb", just = c("left", "top"))

## Hide page guides
bb_pageGuideHide()

}
