% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_annoYaxis.R
\name{bb_annoYaxis}
\alias{bb_annoYaxis}
\title{wrapper function for plotting an y axis in its own viewport}
\usage{
bb_annoYaxis(
  plot,
  params = NULL,
  at = NULL,
  label = TRUE,
  main = TRUE,
  gp = gpar()
)
}
\arguments{
\item{plot}{plot to add axis to}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{at}{a numeric vector of x-value locations for the tick marks}

\item{label}{a logical value indicating whether to draw the labels on the tick marks, or an expression or character vector which specify the labels to use. If not logical, must be the same length as the at argument}

\item{main}{A logical value indicating whether to draw the axis at the left (TRUE) or at the right (FALSE) of the plot}
}
\description{
wrapper function for plotting an y axis in its own viewport
}
