% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_assembly.R
\name{bb_assembly}
\alias{bb_assembly}
\title{Makes a bb_assembly object for alternate TxDB, OrgDb,and BSgenome packages}
\usage{
bb_assembly(
  Genome,
  TxDb,
  OrgDb,
  gene.id.column = "ENTREZID",
  display.column = "SYMBOL",
  BSgenome = NULL
)
}
\arguments{
\item{Genome}{string of genome assembly name}

\item{TxDb}{string of the desired TxDb package}

\item{OrgDb}{string of the desired OrgDb package}

\item{gene.id.column}{string name of OrgDb column that refers to the given TxDb gene ID's}

\item{display.column}{string name of OrgDb column that is the type of gene symbol to be displayed in plots}

\item{BSgenome}{string of desired BSgenome package}
}
\description{
Makes a bb_assembly object for alternate TxDB, OrgDb,and BSgenome packages
}
