% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bb_CasesUSA}
\alias{bb_CasesUSA}
\title{BentoBox example data for COVID-19 cases in the United States}
\format{
a dataframe with 7 columns
\describe{
\item{state}{The associated state in the United States.}
\item{group}{Numeric value describing a group for each \code{state}.}
\item{long}{Longitude value.}
\item{lat}{Latitude value.}
\item{cases}{The cumulative number of COVID-19 cases.}
\item{population}{Numeric value of total \code{state} population.}
\item{cases_100K}{The cumulative number of COVID-19 cases, per 100000 individuals.}
}
}
\source{
COVID-19 case data was downloaded from The COVID Tracking Project
\url{https://covidtracking.com/}. Data was turned into map data with
\link[ggplot2]{map_data}.
}
\usage{
data("bb_CasesUSA")
}
\description{
A data frame of United States map data and COVID-19 cases as of 2021-03-07.
}
\keyword{datasets}
