% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bb_DeathsUSA}
\alias{bb_DeathsUSA}
\title{BentoBox example data for COVID-19 deaths in the United States}
\format{
a dataframe with 5 columns
\describe{
\item{state}{The associated state in the United States.}
\item{group}{Numeric value describing a group for each `state`.}
\item{long}{Longitude value.}
\item{lat}{Latitude value.}
\item{deaths}{The cumulative number of COVID-19 deaths.}
}
}
\source{
COVID-19 death data was downloaded from The COVID Tracking Project
\url{https://covidtracking.com/}. Data was turned into map data with
\link[ggplot2]{map_data}.
}
\usage{
data("bb_DeathsUSA")
}
\description{
A data frame of United States map data and COVID-19 deaths as of 2021-03-07.
}
\keyword{datasets}
