% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_defaultPackages.R
\name{bb_defaultPackages}
\alias{bb_defaultPackages}
\title{Display the default genomic annotation packages associated with a genome build}
\usage{
bb_defaultPackages(Genome)
}
\arguments{
\item{Genome}{String indicating the name of the genome assembly.}
}
\value{
Returns a list of the default data packages for a genome build.
}
\description{
Display the default genomic annotation packages associated with a genome build
}
\examples{
## View default genomic annotation packages associated with "hg19"
bb_defaultPackages(Genome = "hg19")

## View default genomic annotation packages associated with "mm9"
bb_defaultPackages(Genome = "mm9")

}
