% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_genomes.R
\name{bb_genomes}
\alias{bb_genomes}
\title{Display the included available default genome assemblies}
\usage{
bb_genomes()
}
\value{
Returns the included available default genome assemblies
}
\description{
Display the included available default genome assemblies
}
\examples{
bb_genomes()

}
