% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_groupPlots.R
\name{bb_groupPlots}
\alias{bb_groupPlots}
\title{creates a group of grobs}
\usage{
bb_groupPlots(plots)
}
\arguments{
\item{plots}{a list of plot objects returned from BentoBox functions}
}
\value{
Function will return a "bb_group" object
}
\description{
creates a group of grobs
}
