% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bb_gwasData}
\alias{bb_gwasData}
\title{BentoBox example GWAS data}
\format{
a dataframe in BED format with a column representing p-values
\describe{
\item{chrom}{The name of the chromosome on which the genome feature exists.}
\item{start}{The starting position of the feature in the chromosome.}
\item{end}{The ending position of the feature in the chromosome.}
\item{pVal}{The associated p-value of the feature.}
}
}
\usage{
data("bb_gwasData")
}
\description{
A dataset representing GWAS data in the region chr21:28000000-30300000 according to the hg19 genome build.
}
\keyword{datasets}
