% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bb_hicData}
\alias{bb_hicData}
\title{BentoBox example Hi-C data}
\format{
a 3-column data frame in sparse upper triangular format.
}
\usage{
data("bb_hicData")
}
\description{
A dataset containing interaction frequency matrix counts along genomic coordinates in the
region chr21:28000000-30300000 according to the hg19 genome build. This data is from the IMR90
cell line.
}
\references{
Rao SS, Huntley MH, Durand NC, Stamenova EK, Bochkov ID, Robinson JT, Sanborn AL, Machol I, Omer AD, Lander ES, Aiden EL. A 3D map of the human genome at kilobase resolution reveals principles of chromatin looping. Cell. 2014 Dec 18;159(7):1665-80. doi: 10.1016/j.cell.2014.11.021. Epub 2014 Dec 11. Erratum in: Cell. 2015 Jul 30;162(3):687-8. PMID: 25497547; PMCID: PMC5635824.
(\href{https://pubmed.ncbi.nlm.nih.gov/25497547/}{PubMed})
}
\keyword{datasets}
