% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_labelgenome.R
\name{bb_labelgenome}
\alias{bb_labelgenome}
\title{Adds genome coordinates to the axis of a plot}
\usage{
bb_labelgenome(plot, scale = "bp", side = "bottom", fontsize = 10,
  color = "grey")
}
\arguments{
\item{plot}{plot to annotate (saved as output from BentoBox plotting function)}

\item{scale}{scale of the plot; options are "bp", "Kb", or "Mb"}

\item{side}{side of the plot to annotate, options are "bottom", "top", "right", or "left"}

\item{fontsize}{fontsize for labels (in points)}

\item{color}{color for all text/lines}
}
\description{
Adds genome coordinates to the axis of a plot
}
