% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_makepage.R
\name{bb_makePage}
\alias{bb_makePage}
\title{Makes a page for a BentoBox layout}
\usage{
bb_makePage(
  width = 8.5,
  height = 11,
  default.units = "inches",
  showOutline = TRUE,
  showRuler = TRUE,
  xgrid = 0.5,
  ygrid = 0.5
)
}
\arguments{
\item{width}{A numeric specifying width}

\item{height}{A numeric ospecifying height}

\item{default.units}{A string indicating the units of width and height}

\item{showOutline}{TRUE/FALSE indicating whether to draw a black outline around the entire page}

\item{showRuler}{TRUE/FALSE indicating whether to show guiding ruler along top and side of page}

\item{xgrid}{vertical gridlines}

\item{ygrid}{horizontal gridlines}
}
\description{
Makes a page for a BentoBox layout
}
