% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_maptocolors.R
\name{bb_maptocolors}
\alias{bb_maptocolors}
\title{Maps numeric vector to color palette}
\usage{
bb_maptocolors(vec, col, num = 100, range = NULL)
}
\arguments{
\item{vec}{numeric vector to map to color}

\item{col}{color palette to map to}

\item{num}{number of bins of colors}

\item{range}{range of values to map}
}
\description{
Maps numeric vector to color palette
}
