% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pageCreate.R
\name{bb_pageCreate}
\alias{bb_pageCreate}
\title{Create a page for a BentoBox layout}
\usage{
bb_pageCreate(
  width = 8.5,
  height = 11,
  default.units = "inches",
  xgrid = 0.5,
  ygrid = 0.5,
  showOutline = TRUE,
  showRuler = TRUE,
  showUnit = TRUE,
  params = NULL
)
}
\arguments{
\item{width}{A numeric or unit object specifying page width. Default value is \code{width = 8}.}

\item{height}{A numeric or unit object specifying page height. Default value is \code{height = 11}.}

\item{default.units}{A string indicating the default units to use if \code{width} or \code{height} are only given as numerics. Default value is \code{default.units = "inches"}.}

\item{xgrid}{A numeric indicating the increment by which to place vertical gridlines. Default value is \code{xgrid = 0.5}.}

\item{ygrid}{A numeric indicating the increment by which to place horizontal gridlines. Default value is \code{ygrid = 0.5}.}

\item{showOutline}{A logical value indicating whether to draw a black border around the entire page. Default value is \code{showOutline = TRUE}.}

\item{showRuler}{A logical value indicating whether to show guiding rulers along the top and left side of page. Default value is \code{showRuler = TRUE}.}

\item{showUnit}{A logical value indicating whether to show the name of units at the top left of the page. Default value is \code{showUnit = TRUE}.}

\item{params}{An optional \link[BentoBox]{bb_assembly} object containing relevant function parameters.}
}
\description{
Create a page for a BentoBox layout
}
\details{
\code{width} and \code{height} must be specified in the same units.
}
\examples{
## Create a 4-inch wide, 5-inch high BentoBox page
bb_pageCreate(width = 4, height = 4, default.units = "inches")

## Create a 10-cm wide, 12-cm high Bentobox page
bb_pageCreate(width = 10, height = 10, default.units = "cm")

}
