% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pageGuideShow.R
\name{bb_pageGuideShow}
\alias{bb_pageGuideShow}
\title{Reshow guides drawn with \code{bb_pageCreate},
\code{bb_pageGuideHorizontal}, and \code{bb_pageGuideVertical}}
\usage{
bb_pageGuideShow()
}
\value{
None.
}
\description{
Reshow guides drawn with \code{bb_pageCreate},
\code{bb_pageGuideHorizontal}, and \code{bb_pageGuideVertical}
}
\examples{
## Load Hi-C data
library(BentoBoxData)
data("IMR90_HiC_10kb")

## Create a page
bb_pageCreate(width = 3, height = 3, default.units = "inches")

## Add a page guide
bb_pageGuideHorizontal(y = 0.5, default.units = "inches")

## Plot and place Hi-C plot
hicPlot <- bb_plotHicSquare(
    data = IMR90_HiC_10kb, resolution = 10000,
    zrange = c(0, 70),
    chrom = "chr21",
    chromstart = 28000000, chromend = 30300000,
    x = 0.5, y = 0.5, width = 2, height = 2,
    just = c("left", "top"),
    default.units = "inches"
)

## Hide page guides
bb_pageGuideHide()

## Re-show page guides
bb_pageGuideShow()

## Annotate genome label
bb_annoGenomeLabel(
    plot = hicPlot, scale = "Mb", axis = "x",
    x = 0.5, y = 2.53, just = c("left", "top")
)
}
\seealso{
\link[BentoBox]{bb_pageCreate},
\link[BentoBox]{bb_pageGuideHorizontal},
\link[BentoBox]{bb_pageGuideVertical}
}
