% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pageMake.R
\name{bb_pageMake}
\alias{bb_pageMake}
\title{Makes a page for a BentoBox layout}
\usage{
bb_pageMake(
  params = NULL,
  width = 8.5,
  height = 11,
  default.units = "inches",
  showOutline = TRUE,
  showRuler = TRUE,
  showUnit = TRUE,
  xgrid = 0.5,
  ygrid = 0.5
)
}
\arguments{
\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{width}{A numeric specifying width}

\item{height}{A numeric ospecifying height}

\item{default.units}{A string indicating the units of width and height}

\item{showOutline}{TRUE/FALSE indicating whether to draw a black outline around the entire page}

\item{showRuler}{TRUE/FALSE indicating whether to show guiding ruler along top and side of page}

\item{showUnit}{TRUE/FALSE indicating whether to show name of units in top left of page}

\item{xgrid}{vertical gridlines}

\item{ygrid}{horizontal gridlines}
}
\description{
Makes a page for a BentoBox layout
}
