% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pagePlotRemove.R
\name{bb_pagePlotRemove}
\alias{bb_pagePlotRemove}
\title{Remove BentoBox plots and annotations}
\usage{
bb_pagePlotRemove(plot)
}
\arguments{
\item{plot}{BentoBox plot object to be removed, defined by the output of a BentoBox plotting function.}
}
\description{
Remove BentoBox plots and annotations
}
\examples{
## Load Hi-C data
data("bb_hicData")

## Create BentoBox page
bb_pageCreate(width = 4, height = 3.5, default.units = "inches", xgrid = 0, ygrid = 0)

## Plot and place a square Hi-C plot
hicPlot <- bb_plotHicSquare(hicData = bb_hicData, resolution = 10000, zrange = c(0, 70),
                            chrom = "chr21", chromstart = 28000000, chromend = 30300000,
                            x = 0.5, y = 0.5, width = 2.5, height = 2.5,
                            just = c("left", "top"), default.units = "inches")

## Remove square Hi-C plot from page
bb_pagePlotRemove(plot = hicPlot)

}
