% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pageRemovePlot.R
\name{bb_pageRemovePlot}
\alias{bb_pageRemovePlot}
\title{Removes BentoBox plots and annotations}
\usage{
bb_pageRemovePlot(plot)
}
\arguments{
\item{plot}{BentoBox plot to remove, defined by the output of a BentoBox plotting function.}
}
\description{
Removes BentoBox plots and annotations
}
