% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_placeGroup.R
\name{bb_placeGroup}
\alias{bb_placeGroup}
\title{moves and/or rescales a group of grobs}
\usage{
bb_placeGroup(group, x = NULL, y = NULL, width = NULL,
  height = NULL, just = c("left", "top"), draw = T)
}
\arguments{
\item{group}{a "bb_group" object, as outputted from bb_groupPlots}

\item{x}{A unit object specifying x-location.}

\item{y}{A unit object specifying y-location.}

\item{width}{A unit object specifying width.}

\item{height}{A unit object specifying height.}

\item{just}{A string or numeric vector specifying the justification of the plot relative to its (x, y) location}
}
\value{
Function will return an updated "bb_group" object
}
\description{
moves and/or rescales a group of grobs
}
