% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_placePlot.R
\name{bb_placePlot}
\alias{bb_placePlot}
\title{places a plot that has been previously created but not drawn}
\usage{
bb_placePlot(
  plot,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = T
)
}
\arguments{
\item{plot}{plot to be placed}

\item{x}{A unit object specifying x-location.}

\item{y}{A unit object specifying y-location.}

\item{width}{A unit object specifying width.}

\item{height}{A unit object specifying height.}

\item{just}{A string or numeric vector specifying the justification of the plot relative to its (x, y) location}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numeric vectors}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\value{
Function will update dimensions of an input plot and return an updated BentoBox plot object
}
\description{
places a plot that has been previously created but not drawn
}
