% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotApa.R
\name{bb_plotApa}
\alias{bb_plotApa}
\title{plots an Apa plot}
\usage{
bb_plotApa(apa, loopNumber = 1, palette = colorRampPalette(c("white",
  "dark red")), zrange = NULL, x = 1, y = 1, width = 1,
  height = 1, just = c("center"), units = "inches")
}
\arguments{
\item{apa}{path to APA txt file}

\item{loopNumber}{number of DNA loops}

\item{palette}{ColorRamp palette to use for representing interaction scores}

\item{x}{x-coordinate of plot}

\item{y}{y-coordinate of plot}

\item{width}{width of plot}

\item{height}{height of plot}

\item{just}{justification of x and y-coordinates}

\item{units}{units of width, height, x, and y-coordinates}

\item{zrangethe}{range of interaction scores to plot, where extreme values will be set to the max or min}
}
\description{
plots an Apa plot
}
