% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotBase.R
\name{bb_plotBase}
\alias{bb_plotBase}
\title{adds a plot created with baseR to a BentoBox layout}
\usage{
bb_plotBase(plot, x, y, width, height, just = c("left", "top"))
}
\arguments{
\item{plot}{plot formula of base R plotting functions; ex: p <- ~plot(1:10) + abline(v = 2)}

\item{x}{A unit object specifying x-location.}

\item{y}{A unit object specifying y-location.}

\item{width}{A unit object specifying width.}

\item{height}{A unit object specifying height.}

\item{just}{A string or numeric vector specifying the justification of the plot relative to its (x, y) location}
}
\description{
adds a plot created with baseR to a BentoBox layout
}
