% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotBedpeArches.R
\name{bb_plotBedpeArches}
\alias{bb_plotBedpeArches}
\title{plots data stored in BEDPE format in a "ribbon arch" style}
\usage{
bb_plotBedpeArches(
  bedpe,
  chrom,
  params = NULL,
  chromstart = NULL,
  chromend = NULL,
  cutoffArches = FALSE,
  archHeight = NULL,
  style = "2D",
  arch = 5,
  position = "top",
  fillcolor = "lightgrey",
  linecolor = NA,
  colorby = NULL,
  colorbyrange = NULL,
  assembly = "hg19",
  alpha = 0.4,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = T,
  ...
)
}
\arguments{
\item{bedpe}{bed paired end data to be plotted}

\item{chrom}{chromsome of region to be plotted}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{chromstart}{start position}

\item{chromend}{end position}

\item{cutoffArches}{a logical value indicating whether to include arches that get cutoff in the region}

\item{archHeight}{single value or vector specifying the arch heights;
when NULL, all arches will be the same height, filling up the given viewport}

\item{style}{style of arches: "2D" or "3D"}

\item{arch}{arch curvature, determined by number of points along the curve}

\item{position}{position of arches above or below the x-axis: "top" or "bottom"}

\item{fillcolor}{single value, vector, or palette specifying colors of arches}

\item{linecolor}{linecolor of line outlining arches}

\item{colorby}{name of column in bedpe data to scale colors by}

\item{colorbyrange}{the range of values to apply a colorby palette scale to, if colorby values are numeric}

\item{assembly}{desired genome assembly}

\item{alpha}{transparency of arches}

\item{x}{A numeric or unit object specifying x-location}

\item{y}{A numeric or unit object specifying y-location}

\item{width}{A numeric or unit object specifying width}

\item{height}{A numeric or unit object specifying height}

\item{just}{A string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numerics}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\value{
Function will plot a ribbon arches plot and return a bb_arches object
}
\description{
plots data stored in BEDPE format in a "ribbon arch" style
}
