% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotCircle.R
\name{bb_plotCircle}
\alias{bb_plotCircle}
\title{Plot a circle within a BentoBox layout}
\usage{
bb_plotCircle(r, x, y, default.units = "inches")
}
\arguments{
\item{r}{A numeric vector or unit object specifying radii.}

\item{x}{A numeric vector or unit object specifying circle x-locations relative to center.}

\item{y}{A numeric vector or unit object specifying circle y-locations relative to center.}

\item{default.units}{A string indicating the default units to use if \code{r}, \code{x}, or \code{y} are only given as numerics or numeric vectors. Default value is \code{default.units = "inches"}.}

\item{linecolor}{A character value specifying circle line color. Default value is \code{linecolor = "black"}.}

\item{lwd}{A numeric specifying circle line width. Default value is \code{lwd = 1}.}

\item{lty}{A numeric specifying circle line type. Default value is \code{lty = 1}.}

\item{fill}{A character value specifying circle fill color. Default value is \code{fill = NA}.}

\item{alpha}{Numeric value specifying color transparency. Default value is \code{alpha = 1}.}

\item{params}{An optional \link[BentoBox]{bb_assembly} object containing relevant function parameters.}

\item{...}{Additional grid graphical parameters. See \link[grid]{gpar}.}
}
\value{
Returns a \code{bb_circle} object containing relevant placement and \link[grid]{grob} information.
}
\description{
Plot a circle within a BentoBox layout
}
\examples{
## Create a BentoBox page
bb_pageCreate(width = 2, height = 2, default.units = "inches", xgrid = 0, ygrid = 0)

## Plot a circle
bb_plotCircle(r = 0.3, x = 1, y = 1.2, default.units = "inches")

}
\seealso{
\link[grid]{grid.circle}
}
