% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotgenes.R
\name{bb_plotGenes}
\alias{bb_plotGenes}
\title{plots a gene track for a specified region}
\usage{
bb_plotGenes(
  assembly = "hg19",
  chrom,
  chromstart,
  chromend,
  fontcolors = c("#2929ff", "#ff3434"),
  strandcolors = c("#8a8aff", "#ff7e7e"),
  fontsize = 8,
  strandLabels = T,
  x = NULL,
  y = NULL,
  width = NULL,
  height = unit(0.6, "inches"),
  just = c("left", "top"),
  default.units = "inches",
  draw = T
)
}
\arguments{
\item{assembly}{desired genome assembly}

\item{chrom}{chromsome of region to be plotted}

\item{chromstart}{start position}

\item{chromend}{end position}

\item{fontcolors}{a vector indicating the font colors for the plus strand and minus strand gene labels}

\item{strandcolors}{a vector indicating the strand colors for the plus strand and minus strand}

\item{fontsize}{the size of gene label text (in points)}

\item{strandLabels}{A logical value indicating whether to include +/- strand labels}

\item{x}{A numeric or unit object specifying x-location}

\item{y}{A numeric or unit object specifying y-location}

\item{width}{A numeric or unit object specifying width}

\item{height}{A numeric or unit object specifying height}

\item{just}{string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numerics}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\description{
plots a gene track for a specified region
}
