% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotHic.R
\name{bb_plotHic}
\alias{bb_plotHic}
\title{plots HiC interaction matrix}
\usage{
bb_plotHic(hic, chrom = 8, chromstart = 133600000,
  chromend = 134800000, half = "both", resolution = 10000,
  zrange = NULL, palette = colorRampPalette(c("white", "dark red")),
  width = 3, height = 3, x = 1, y = 1, just = c("left", "top"),
  units = "inches", altchrom = NULL, altchromstart = NULL,
  altchromend = NULL, althalf = NULL, norm = "KR", ...)
}
\arguments{
\item{hic}{path to .hic file or 3 column dataframe of counts}

\item{chrom}{chromosome of region to be plotted}

\item{chromstart}{chromosome start of region to be plotted}

\item{chromend}{chromosome end of region to be plotted}

\item{half}{what sides of square plot; options are "both", top", or "bottom"}

\item{resolution}{the width in bp of each pixel; options are 2500000, 1000000, 500000, 250000, 100000, 50000, 25000, 10000, or 5000}

\item{zrange}{the range of interaction scores to plot, where extreme values will be set to the max or min}

\item{palette}{ColorRamp palette to use for representing interaction scores}

\item{width}{width of plot in specified units}

\item{height}{height of plot in specified units}

\item{x}{x-coordinate of where to place plot based on just}

\item{y}{y-coordinate of where to place plot based on just}

\item{just}{a string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{units}{units of height, width, and x and y location of the plot}

\item{altchrom}{alternate chromosome for off-diagonal plotting or interchromosomal plotting}

\item{altchromstart}{alternate chromosome start for off-diagonal plotting or interchromosomal plotting}

\item{altchromend}{alternate chromosome end for off-diagonal plotting or interchromosomal plotting}

\item{althalf}{if plotting altchrom region, which side off diagonal to plot; options are "top" or "bottom"}

\item{norm}{if giving .hic file, hic data normalization; options are "NONE", "VC", "VC_SQRT", and "KR"}
}
\value{
Function will plot a HiC interaction matrix and return a bb_hicPlot object
}
\description{
plots HiC interaction matrix
}
\examples{

}
\author{
Nicole Kramer
}
