% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotHicSquare.R
\name{bb_plotHicSquare}
\alias{bb_plotHicSquare}
\title{plots HiC interaction matrix}
\usage{
bb_plotHicSquare(
  hic,
  chrom,
  params = NULL,
  chromstart = NULL,
  chromend = NULL,
  half = "both",
  resolution = "auto",
  zrange = NULL,
  palette = colorRampPalette(c("white", "dark red")),
  assembly = "hg19",
  width = NULL,
  height = NULL,
  x = NULL,
  y = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  altchrom = NULL,
  altchromstart = NULL,
  altchromend = NULL,
  althalf = NULL,
  norm = "KR"
)
}
\arguments{
\item{hic}{path to .hic file or 3 column dataframe of counts}

\item{chrom}{chromosome of region to be plotted}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{chromstart}{start position}

\item{chromend}{end position}

\item{half}{what sides of square plot; options are "both", top", or "bottom"}

\item{resolution}{the width in bp of each pixel; for hic files, "auto" will attempt to choose a resolution based on the size of the region; for
dataframes, "auto" will attempt to detect the resolution the dataframe contains}

\item{zrange}{the range of interaction scores to plot, where extreme values will be set to the max or min}

\item{palette}{ColorRamp palette to use for representing interaction scores}

\item{assembly}{default genome assembly as a string or a bb_assembly object}

\item{width}{A numeric or unit object specifying width}

\item{height}{A numeric or unit object specifying height}

\item{x}{A numeric or unit object specifying x-location}

\item{y}{A numeric or unit object specifying y-location}

\item{just}{a string or numeric vector specifying the justification of the viewport relative to its (x, y) location}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numeric vectors}

\item{draw}{A logical value indicating whether graphics output should be produced}

\item{altchrom}{alternate chromosome for off-diagonal plotting or interchromosomal plotting}

\item{altchromstart}{alternate chromosome start for off-diagonal plotting or interchromosomal plotting}

\item{altchromend}{alternate chromosome end for off-diagonal plotting or interchromosomal plotting}

\item{althalf}{if plotting altchrom region, which side off diagonal to plot; options are "top" or "bottom"}

\item{norm}{if giving .hic file, hic data normalization; must be found in hic file}
}
\value{
Function will plot a HiC interaction matrix and return a bb_hic object
}
\description{
plots HiC interaction matrix
}
\examples{

}
