% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotIdeogram.R
\name{bb_plotIdeogram}
\alias{bb_plotIdeogram}
\title{Plot a chromosome ideogram with or without cytobands}
\usage{
bb_plotIdeogram(
  chrom,
  assembly = "hg19",
  orientation = "h",
  showBands = TRUE,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  params = NULL,
  ...
)
}
\arguments{
\item{chrom}{Chromosome to be plotted, as a string.}

\item{assembly}{Default genome assembly as a string or a \link[BentoBox]{bb_assembly} object. Default value is \code{assembly = "hg19"}.}

\item{orientation}{Character value indicating the orientation of the ideogram. Default value is \code{orientation = "h"}. Options are:
\itemize{
\item{\code{"v"}: }{Vertical ideogram orientation.}
\item{\code{"h"}: }{Horizontal ideogram orientation.}
}}

\item{showBands}{Logical value indicating whether to draw colored cytobands within ideogram. Default value is \code{showBands = TRUE}.}

\item{x}{A numeric or unit object specifying ideogram x-location.}

\item{y}{A numeric or unit object specifying ideogram y-location.}

\item{width}{A numeric or unit object specifying ideogram width.}

\item{height}{A numeric or unit object specifying ideogram height.}

\item{just}{Justification of ideogram relative to its (x, y) location. If there are two values, the first value specifies horizontal justification and the second value specifies vertical justification.
Possible string values are: \code{"left"}, \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"}, and \code{"top"}. Default value is \code{just = c("left", "top")}.}

\item{default.units}{A string indicating the default units to use if \code{x}, \code{y}, \code{width}, or \code{height} are only given as numerics. Default value is \code{default.units = "inches"}.}

\item{draw}{A logical value indicating whether graphics output should be produced. Default value is \code{draw = TRUE}.}

\item{params}{An optional \link[BentoBox]{bb_assembly} object containing relevant function parameters.}

\item{...}{Additional grid graphical parameters. See \link[grid]{gpar}.}
}
\value{
Returns a \code{bb_ideogram} object containing relevant genomic region, placement, and \link[grid]{grob} information.
}
\description{
Plot a chromosome ideogram with or without cytobands
}
\details{
This function can be used to quickly plot an ideogram by ignoring plot placement parameters:
\preformatted{
bb_plotIdeogram(chrom)
}
An ideogram can be placed on a BentoBox coordinate page by providing plot placement parameters:
\preformatted{
bb_plotIdeogram(chrom,
                x, y, width, height, just = c("left", "top"),
                default.units = "inches")
}

Giemsa stain band data from the UCSC Genome Browser is included with BentoBox.
}
\examples{
## Load Giemsa stain band information and genomic annotation data for hg19 genome assembly
library("TxDb.Hsapiens.UCSC.hg19.knownGene")
data("cytoBand.Hsapiens.UCSC.hg19")

## Plot ideogram filling up entire graphic device
bb_plotIdeogram(chrom = "chr2", assembly = "hg19")

## Plot and place ideogram on a BentoBox page
bb_pageCreate(width = 4, height = 1.5, default.units = "inches")
ideogramPlot <- bb_plotIdeogram(chrom = "chr2", assembly = "hg19",
                                x = 0.25, y = 0.25, width = 3.5, height = 0.5,
                               just = c("left", "top"), default.units = "inches")

## Annotate genome label
bb_annoGenomeLabel(plot = ideogramPlot, x = 0.25, y = 1, scale = "Mb")

## Hide page guides
bb_pageGuideHide()

}
