% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotKaryogram.R
\name{bb_plotKaryogram}
\alias{bb_plotKaryogram}
\title{plots and highlights a chromosome with its cytobands}
\usage{
bb_plotKaryogram(
  assembly = "hg19",
  chrom,
  start = NULL,
  end = NULL,
  highlightCol = "red",
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  ...
)
}
\arguments{
\item{assembly}{desired genome assembly}

\item{chrom}{chromsome to plot}

\item{start}{highlight start}

\item{end}{highlight end}

\item{highlightCol}{fillcolor and linecolor for highlight box}

\item{x}{A numeric or unit object specifying x-location}

\item{y}{A numeric or unit object specifying y-location}

\item{width}{A numeric or unit object specifying width}

\item{height}{A numeric or unit object specifying height}

\item{just}{A string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numerics}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\value{
Function will return a bb_karyogram object
}
\description{
plots and highlights a chromosome with its cytobands
}
