% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotLegend.R
\name{bb_plotLegend}
\alias{bb_plotLegend}
\title{plots a legend}
\usage{
bb_plotLegend(
  legend,
  params = NULL,
  orientation = "v",
  fillcolor = NULL,
  pch = NULL,
  lty = NULL,
  title = NULL,
  fontsize = 10,
  border = TRUE,
  bg = NA,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  ...
)
}
\arguments{
\item{legend}{single value or vector of legend text#'}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{orientation}{"v" (vertical) or "h" (horizontal) orientation}

\item{fillcolor}{if specified, this argument will produce boxes filled with the specified colors to appear beside the legend text}

\item{pch}{if specified, this argument will produce symbols to appear beside the legend text}

\item{lty}{if specified, this argument will produce line types to appear beside the legend text}

\item{title}{legend title}

\item{border}{option to add border around legend}

\item{bg}{background color of legend}

\item{x}{A numeric vector or unit object specifying x-location}

\item{y}{A numeric vector or unit object specifying y-location}

\item{width}{A numeric vector or unit object specifying width}

\item{height}{A numeric vector or unit object specifying height}

\item{just}{justification of scale viewport}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numeric vectors}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\description{
plots a legend
}
