% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotManhattan.R
\name{bb_plotManhattan}
\alias{bb_plotManhattan}
\title{plots a Manhattan plot}
\usage{
bb_plotManhattan(bedfile, chrom, chromstart, chromend, col, ymax = 1,
  ylim = NULL, x, y, width, height, just, units, ...)
}
\arguments{
\item{bedfile}{bedfile for Manhattan plot}

\item{chrom}{chromosome of region to be plotted}

\item{chromstart}{chromstart of region to be plotted}

\item{chromend}{chromend of region to be plotted}

\item{col}{single colors, vector of colors, or color palette for coloring points}

\item{ymax}{fraction of max y value to set as height of plot}

\item{ylim}{numeric vector of length 2, giving the y coordinates range}

\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{width}{width of plot}

\item{height}{height of plot}

\item{just}{justification of plot}

\item{units}{units of width, height, x, and y-coordinates}
}
\description{
plots a Manhattan plot
}
