% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotPileup.R
\name{bb_plotPileup}
\alias{bb_plotPileup}
\title{plots data stored in BED format in a pileup style}
\usage{
bb_plotPileup(
  bed,
  chrom,
  params = NULL,
  chromstart = NULL,
  chromend = NULL,
  assembly = "hg19",
  fillcolor = "black",
  linecolor = NA,
  colorby = NULL,
  colorbyrange = NULL,
  strandSplit = FALSE,
  boxHeight = unit(2, "mm"),
  spaceHeight = 0.3,
  spaceWidth = 0.02,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  ...
)
}
\arguments{
\item{bed}{genomic data to be plotted; can be any data in BED format, a .bam file where a bam index file (.bam.bai) is in the same directory,  or a GRanges}

\item{chrom}{chromosome of region to be plotted}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{chromstart}{start position}

\item{chromend}{end position}

\item{assembly}{desired genome assembly}

\item{fillcolor}{single value, vector, or palette specifying colors of pileup elements}

\item{linecolor}{linecolor of line outlining pileup elements}

\item{colorby}{name of column in bed data to scale colors by}

\item{colorbyrange}{the range of values to apply a colorby palette scale to, if colorby values are numeric}

\item{strandSplit}{logical indicating whether plus and minus-stranded elements should be separated}

\item{boxHeight}{A numeric or unit object specifying height of pileup element boxes}

\item{spaceHeight}{height of spacing between pileup element boxes, as a fraction of boxHeight}

\item{spaceWidth}{width of minimum spacing between pileup element boxes, as a fraction of the plot's genomic range}

\item{x}{A numeric or unit object specifying x-location}

\item{y}{A numeric or unit object specifying y-location}

\item{width}{A numeric or unit object specifying width}

\item{height}{A numeric or unit object specifying height}

\item{just}{A string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numerics}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\value{
Function will plot a pileup style plot and return a bb_pileup object
}
\description{
plots data stored in BED format in a pileup style
}
