% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotPolygon.R
\name{bb_plotPolygon}
\alias{bb_plotPolygon}
\title{wrapper to draw a polygonGrob based on BentoBox page coordinates and units}
\usage{
bb_plotPolygon(
  x,
  y,
  id = NULL,
  id.lengths = NULL,
  params = NULL,
  linecolor = "black",
  fill = NA,
  lwd = 1,
  lty = 1,
  alpha = 0,
  default.units = "inches",
  ...
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-location}

\item{y}{A numeric vector or unit object specifying y-location}

\item{id}{A numeric vector used to separate locations in x and y into multiple polygons. All locations with the same id belong to the same polygon.}

\item{id.lengths}{A numeric vector used to separate locations in x and y into multiple polygons. Specifies consecutive blocks of locations which make up separate polygons.}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{linecolor}{line color}

\item{fill}{fill color}

\item{lwd}{line width}

\item{lty}{line type}

\item{alpha}{color transparency}

\item{default.units}{A string indicating the default units to use if x or y are only given as numeric vectors}
}
\description{
wrapper to draw a polygonGrob based on BentoBox page coordinates and units
}
