% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotRect.R
\name{bb_plotRect}
\alias{bb_plotRect}
\title{wrapper to draw a rectGrob based on BentoBox page coordinates and units}
\usage{
bb_plotRect(
  x,
  y,
  width,
  height,
  just = "center",
  params = NULL,
  linecolor = "black",
  fill = NA,
  lwd = 1,
  lty = 1,
  alpha = 1,
  default.units = "inches",
  ...
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-location}

\item{y}{A numeric vector or unit object specifying y-location}

\item{width}{A numeric vector or unit object specifying width}

\item{height}{A numeric vector or unit object specifying height}

\item{just}{The justification of the rectangle relative to its (x,y) location}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{linecolor}{line color}

\item{fill}{fill color}

\item{lwd}{line width}

\item{lty}{line type}

\item{alpha}{color transparency}

\item{default.units}{A string indicating the default units to use if x or y are only given as numeric vectors}
}
\description{
wrapper to draw a rectGrob based on BentoBox page coordinates and units
}
