% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotSegments.R
\name{bb_plotSegments}
\alias{bb_plotSegments}
\title{wrapper to draw a segmentsGrob based on BentoBox page coordinates and units}
\usage{
bb_plotSegments(
  x0,
  y0,
  x1,
  y1,
  params = NULL,
  arrow = NULL,
  linecolor = "black",
  lwd = 1,
  lty = 1,
  lineend = "butt",
  linejoin = "mitre",
  alpha = 1,
  default.units = "inches",
  ...
)
}
\arguments{
\item{x0}{Numeric indicating the starting x-values of the line segments}

\item{y0}{Numeric indicating the starting y-values of the line segments}

\item{x1}{Numeric indicating the stopping x-values of the line segments}

\item{y1}{Numeric indicating the stopping y-values of the line segments}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{arrow}{A list describing arrow heads to place at either end of the line segments, as produced by the arrow function}

\item{linecolor}{line color}

\item{lwd}{line width}

\item{lty}{line type}

\item{lineend}{Line end style (round, butt, square)}

\item{linejoin}{Line join style (round, mitre, bevel)}

\item{alpha}{color transparency}

\item{default.units}{A string indicating the default units to use if x or y are only given as numeric vectors}
}
\description{
wrapper to draw a segmentsGrob based on BentoBox page coordinates and units
}
