% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotSignal.R
\name{bb_plotSignal}
\alias{bb_plotSignal}
\title{plots signal track data}
\usage{
bb_plotSignal(signal, chrom, chromstart, chromend, range = NULL,
  linecolor = "grey", lwd = 1, yaxis = FALSE, binSize = NA,
  binCap = TRUE, fill = FALSE, fillcolor = NA, transparency = NA,
  ymax = 1.04, width = 3.25, height = 0.625, x = 4.25, y = 5.5,
  units = "inches", just = c("left", "top"), ...)
}
\arguments{
\item{signal}{signal track data to be plotted (bigwig file, bedgraph format dataframe, or bedgraph)}

\item{chrom}{chromosome of region to be plotted}

\item{chromstart}{start position}

\item{chromend}{end position}

\item{range}{y-range to plot (c(min, max))}

\item{linecolor}{color of line outlining signal track}

\item{lwd}{linewidth of line outlining signal track}

\item{binSize}{the length of each bin in bp}

\item{binCap}{TRUE/FALSE whether the function will limit the number of bins to 8,000}

\item{ymax}{fraction of max y value to set as height of plot}

\item{width}{width of entire plot, in specified units}

\item{height}{height of entire plot, in specified units}

\item{x}{x-coordinate of where to place plot based on just}

\item{y}{y-coordinate of where to place plot based on just}

\item{units}{units of width, height, x, and y of plot}

\item{just}{a string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"



#' @export}

\item{addscale}{TRUE/FALSE whether to add a y-axis}
}
\description{
plots signal track data
}
