% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotText.R
\name{bb_plotText}
\alias{bb_plotText}
\title{wrapper to draw a textGrob based on BentoBox page coordinates and units}
\usage{
bb_plotText(
  label,
  x,
  y,
  params = NULL,
  just = "center",
  fontcolor = "black",
  fontsize = 12,
  rot = 0,
  check.overlap = FALSE,
  default.units = "inches",
  ...
)
}
\arguments{
\item{label}{character or expression}

\item{x}{A numeric vector or unit object specifying x-location}

\item{y}{A numeric vector or unit object specifying y-location}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{just}{justification of text relative to its (x, y) location}

\item{fontcolor}{fontcolor}

\item{fontsize}{the size of text (in points)}

\item{rot}{the angle to rotate the song}

\item{check.overlap}{A logical value to indicate whether to check for and omit overlapping text.}

\item{default.units}{A string indicating the default units to use if x or y are only given as numeric vectors}
}
\description{
wrapper to draw a textGrob based on BentoBox page coordinates and units
}
