% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotTranscripts.R
\name{bb_plotTranscripts}
\alias{bb_plotTranscripts}
\title{plots gene transcripts in a pileup style}
\usage{
bb_plotTranscripts(
  chrom,
  params = NULL,
  assembly = "hg19",
  chromstart = NULL,
  chromend = NULL,
  boxHeight = unit(2, "mm"),
  spaceHeight = 0.3,
  spaceWidth = 0.02,
  fillcolor = c("#8a8aff", "#ff7e7e"),
  colorbyStrand = TRUE,
  labels = "transcript",
  fontsize = 8,
  strandSplit = FALSE,
  stroke = 0.1,
  bg = NA,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE
)
}
\arguments{
\item{chrom}{chromsome of region to be plotted}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{assembly}{desired genome assembly}

\item{chromstart}{start position}

\item{chromend}{end position}

\item{boxHeight}{A numeric or unit object specifying total height of each transcript}

\item{spaceHeight}{height of spacing between transcripts, as a fraction of boxHeight}

\item{spaceWidth}{width of minimum spacing between transcripts, as a fraction of the plot's genomic range}

\item{fillcolor}{single value or vector specifying colors of transcripts}

\item{colorbyStrand}{a logical value indicating whether to color strands by the first two colors in a fillcolor vector}

\item{labels}{how labels should be determined; options are NULL for no labels, "transcript" for transcript names, "gene" for gene names,
or "both" for combined transcript and gene names}

\item{fontsize}{the size of gene label text (in points)}

\item{strandSplit}{logical indicating whether plus and minus-stranded elements should be separated}

\item{stroke}{numerical value indicating the stroke width for transcript body outlines}

\item{bg}{background color}

\item{x}{A numeric or unit object specifying x-location}

\item{y}{A numeric or unit object specifying y-location}

\item{width}{A numeric or unit object specifying width}

\item{height}{A numeric or unit object specifying height}

\item{just}{string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numerics}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\value{
Function will plot a pileup of transcripts and return a bb_transcripts object
}
\description{
plots gene transcripts in a pileup style
}
