% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotTriangleHic.R
\name{bb_plotTriangleHic}
\alias{bb_plotTriangleHic}
\title{plots HiC interaction matrix in triangular format}
\usage{
bb_plotTriangleHic(
  hic,
  chrom,
  chromstart = NULL,
  chromend = NULL,
  resolution = "auto",
  zrange = NULL,
  palette = colorRampPalette(c("white", "dark red")),
  assembly = "hg19",
  width = NULL,
  height = NULL,
  x = NULL,
  y = NULL,
  just = c("left", "top"),
  norm = "KR",
  default.units = "inches",
  draw = T,
  ...
)
}
\arguments{
\item{hic}{path to .hic file or 3 column dataframe of counts}

\item{chrom}{chromosome of region to be plotted, based on build (i.e. for hg19 just a number, for hg38 string like "chr8")}

\item{chromstart}{chromosome start of region to be plotted}

\item{chromend}{chromosome end of region to be plotted}

\item{resolution}{the width in bp of each pixel; for hic files, "auto" will attempt to choose a resolution based on the size of the region; for
dataframes, "auto" will attempt to detect the resolution the dataframe contains}

\item{zrange}{the range of interaction scores to plot, where extreme values will be set to the max or min}

\item{palette}{ColorRamp palette to use for representing interaction scores}

\item{assembly}{desired genome assembly}

\item{width}{A numeric or unit object specifying the bottom width of the triangle}

\item{height}{A numeric or unit object specifying the height of the triangle}

\item{x}{A numeric or unit object specifying x-location}

\item{y}{A numeric or unit object specifying y-location}

\item{just}{a string or numeric vector specifying the justification of the viewport relative to its (x, y) location}

\item{norm}{if giving .hic file, hic data normalization; must be found in hic file}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numerics}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\description{
plots HiC interaction matrix in triangular format
}
