% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_readHic.R
\name{bb_readHic}
\alias{bb_readHic}
\title{extracts HiC data from .hic file using Straw}
\usage{
bb_readHic(
  hic,
  chrom,
  chromstart = NULL,
  chromend = NULL,
  resolution = 10000,
  zrange = NULL,
  norm = "KR",
  res_scale = "BP",
  altchrom = NULL,
  altchromstart = NULL,
  altchromend = NULL
)
}
\arguments{
\item{hic}{path to .hic file}

\item{chrom}{if not alternative chromosome, chromosome of desired region}

\item{chromstart}{chromosome start position of chrom}

\item{chromend}{chromosome end position of chrom}

\item{resolution}{the width of each pixel}

\item{zrange}{the range of interaction scores to plot, where extreme values will be set to the max or min; if null, zrange will be set to (0, max(data))}

\item{norm}{hic data normalization; must be found in hic file}

\item{res_scale}{resolution scale; options are "BP" and "FRAG"}

\item{altchrom}{if looking at region between two different chromosomes, this is the specified alternative chromosome}

\item{altchromstart}{if looking at region between two different chromosomes, start position of altchrom}

\item{altchromend}{if looking at region between two different chromsomes, end position of altchrom}
}
\description{
extracts HiC data from .hic file using Straw
}
