% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_zoomHic.R
\name{bb_zoomHic}
\alias{bb_zoomHic}
\title{zooms and plots region of Hi-C plot}
\usage{
bb_zoomHic(hic, chromstart, chromend, altchromstart = NULL,
  altchromend = NULL, x, y, width, height, units, just, lty = "solid",
  lwd = 1, col = "black", lineSide = "right")
}
\arguments{
\item{hic}{hic plot to zoom inon}

\item{chromstart}{chromosome start of region to be zoomed}

\item{chromend}{chromosome end of region to be zoomed}

\item{x}{x-coordinate of where to place zoom plot based on just}

\item{y}{y-coordinate of where to place zoom plot based on just}

\item{width}{width of zoom plot in specified units}

\item{height}{height of zoom plot in specified units}

\item{units}{units of height, width, and x and y location of the zoom plot}

\item{just}{a string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{lty}{line type of zoom box; options are "solid" (1), "dashed" (2), "dotted" (3), "dotdash" (4), "longdash" (5), or "twodash" (6)}

\item{lwd}{line width of zoom box}

\item{col}{line color of zoom box}

\item{lineSide}{side of zoom box from which to draw lines; options are NULL, "right", "left", "top", or "bottom"}
}
\description{
zooms and plots region of Hi-C plot
}
