% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbDefaultPackages.R
\name{bbDefaultPackages}
\alias{bbDefaultPackages}
\title{Display the default genomic annotation packages associated with a
genome build}
\usage{
bbDefaultPackages(Genome)
}
\arguments{
\item{Genome}{String indicating the name of the genome assembly.}
}
\value{
Returns a list of the default data packages for a genome build.
}
\description{
Display the default genomic annotation packages associated with a
genome build
}
\examples{
## View default genomic annotation packages associated with "hg19"
bbDefaultPackages(Genome = "hg19")

## View default genomic annotation packages associated with "mm9"
bbDefaultPackages(Genome = "mm9")
}
