% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbGenomes.R
\name{bbGenomes}
\alias{bbGenomes}
\title{Display the included available default genome assemblies}
\usage{
bbGenomes()
}
\value{
Returns the included available default genome assemblies
}
\description{
Display the included available default genome assemblies
}
\examples{
bbGenomes()
}
