% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbPageGuideHide.R
\name{bbPageGuideHide}
\alias{bbPageGuideHide}
\title{Remove guides from a BentoBox page}
\usage{
bbPageGuideHide()
}
\value{
None.
}
\description{
Remove guides from a BentoBox page
}
\examples{
## Make a BentoBox page
bbPageCreate(width = 7, height = 4, default.units = "inches")

## Hide page guides
bbPageGuideHide()
}
