% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultPackages.R
\name{defaultPackages}
\alias{defaultPackages}
\title{Display the default genomic annotation packages associated with a
genome build}
\usage{
defaultPackages(Genome)
}
\arguments{
\item{Genome}{String indicating the name of the genome assembly.}
}
\value{
Returns a list of the default data packages for a genome build.
}
\description{
Display the default genomic annotation packages associated with a
genome build
}
\examples{
## View default genomic annotation packages associated with "hg19"
defaultPackages(Genome = "hg19")

## View default genomic annotation packages associated with "mm9"
defaultPackages(Genome = "mm9")
}
