% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCool.R
\name{.coolAutoResolution}
\alias{.coolAutoResolution}
\title{Determine best resolution for size of region for .(m)cool files}
\usage{
.coolAutoResolution(file, chromstart, chromend)
}
\arguments{
\item{file}{Path to .(m)cool file}

\item{chromstart}{Chromstart of region}

\item{chromend}{Chromend of region}
}
\description{
Determine best resolution for size of region for .(m)cool files
}

\value{
A numeric value specifying the automatically determined resolution in base pairs.
}
