% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractHiC.R
\name{extractHiC}
\alias{extractHiC}
\title{extracts HiC data from .hic file using Straw}
\usage{
extractHiC(hic, format, chrom, chromstart = NULL, chromend = NULL,
  resolution, zrange = NULL, norm = "NONE", resscale = "BP",
  altchrom = NULL, altchromstart = NULL, altchromend = NULL)
}
\arguments{
\item{hic}{path to .hic file}

\item{format}{format of data wanted, which will depend on what kind of plot is ultimately desired; options are "sparse" and "full"}

\item{chrom}{if not alternative chromosome, chromosome of desired region}

\item{chromstart}{chromosome start position of chrom, in bp}

\item{chromend}{chromosome end position of chrom, in bp}

\item{resolution}{the width in bp of each pixel}

\item{zrange}{the range of interaction scores to plot, where extreme values will be set to the max or min}

\item{norm}{hic data normalization; options are "NONE", "VC", "VC_SQRT", and "KR"}

\item{resscale}{scale of normalization; options are "BP" and "FRAG"}

\item{altchrom}{if looking at region between two different chromosomes, this is the specified alternative chromsome}

\item{altchromstart}{if looking at region between two different chromosomes, start position of altchrom}

\item{altchromend}{if looking at region between two different chromsomes, end position of altchrom}
}
\description{
extracts HiC data from .hic file using Straw
}
