% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomes.R
\name{genomes}
\alias{genomes}
\title{Display the included available default genome assemblies}
\usage{
genomes()
}
\value{
Returns the included available default genome assemblies
}
\description{
Display the included available default genome assemblies
}
\examples{
genomes()
}
