% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageGuideHide.R
\name{pageGuideHide}
\alias{pageGuideHide}
\title{Remove guides from a plotgardener page}
\usage{
pageGuideHide()
}
\value{
None.
}
\description{
Remove guides from a plotgardener page
}
\examples{
## Make a page
pageCreate(width = 7, height = 4, default.units = "inches")

## Hide page guides
pageGuideHide()
}
