% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgardener.R
\docType{package}
\name{plotgardener}
\alias{plotgardener}
\alias{plotgardener-package}
\title{plotgardener: Coordinate-based Genomic Visualization Package for R}
\description{
plotgardener is a coordinate-based genomic visualization package for R.
It grants users the ability to programmatically produce complex,
multi-paneled figures. Tailored for genomics, plotgardener allows users
to visualize large complex genomic datasets and
provides exquisite control over how plots are placed and arranged on
a page.
}
\seealso{
Useful links:
\itemize{
    \item \url{https://phanstiellab.github.io/plotgardener}
    \item \url{https://github.com/PhanstielLab/plotgardener}
}
}
\author{
\strong{Maintainer}: Nicole Kramer \email{nekramer@live.unc.edu}
(\href{https://orcid.org/0000-0001-9617-9671}{ORCID})

Authors:
\itemize{
    \item Eric S. Davis \email{esdavis@live.unc.edu}
    (\href{https://orcid.org/0000-0003-4051-3217}{ORCID})
    \item Craig Wenger \email{craig.wenger@gmail.com}
    (\href{https://orcid.org/0000-0002-7361-8456}{ORCID})
    \item Douglas H. Phanstiel \email{douglas_phanstiel@med.unc.edu}
    [copyright holder]
}

Other contributors:
\itemize{
    \item Sarah Parker \email{sarmae@live.unc.edu} [contributor]
    \item Erika Deoudes \email{ed@erikadudes.com} [artist]
    \item Michael Love \email{milove@email.unc.edu} [contributor]
}
}
