% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenes.R
\name{plotGenes}
\alias{plotGenes}
\title{Plot a gene track for a specified genomic region}
\usage{
plotGenes(
    chrom,
    chromstart = NULL,
    chromend = NULL,
    assembly = "hg38",
    fontsize = 8,
    fontcolor = c("#669fd9", "#abcc8e"),
    fill = c("#669fd9", "#abcc8e"),
    geneOrder = NULL,
    geneHighlights = NULL,
    geneBackground = "grey",
    strandLabels = TRUE,
    stroke = 0.1,
    bg = NA,
    x = NULL,
    y = NULL,
    width = NULL,
    height = unit(0.6, "inches"),
    just = c("left", "top"),
    default.units = "inches",
    draw = TRUE,
    params = NULL
)
}
\arguments{
\item{chrom}{string Chromosome of region to be plotted, as a string. Default value is \code{chrom = NULL}. Options = DNE.}

\item{chromstart}{int start position on chromosome to be plotted. Default value is \code{chromstart = NULL}. Options = DNE.}

\item{chromend}{int end position on chromosome to be plotted. Default value is \code{chromend = NULL}. Options = DNE.}

\item{assembly}{string Default genome assembly as a string or a
\link[plotgardener]{assembly} object.
Default value is \code{assembly = "hg38"}. Options = DNE.}

\item{fontsize}{numeric Specifying text fontsize in points.
Default value is \code{fontsize = 8}. Options include \code{any numeric}}

\item{fontcolor}{string Vector of length 2 indicating
the fontcolors for the plus strand and minus strand gene labels.
The first value will color the plus strand gene labels and
the second value will color the minus strand gene labels.
Default value is \code{fontcolor = c("#669fd9", "#abcc8e")}. Options = hexademical vector}

\item{fill}{vector A character value or vector of length 2 indicating the
strand fill colors for the plus strand and minus strand plot elements.
The first value will color the plus strand plot elements and
the second label will color the minus strand plot elements.
Default value is \code{fill = c("#669fd9", "#abcc8e")}. Options = hexcode color}

\item{geneOrder}{vector An ordered character vector of gene names to
prioritize when labeling genes. Default value is \code{geneOrder = NULL}.}

\item{geneHighlights}{dataframe A two-column dataframe with a column named "gene"
containing gene names as strings to highlight and a named column "color"
containing corresponding highlight colors. Default value is \code{geneHighlights = NULL}.}

\item{geneBackground}{string If \code{geneHighlights} is given, a character
value indicating the color for genes that are not highlighted. Default value is \code{geneBackground = "grey"}}

\item{strandLabels}{bool A logical value indicating whether to include
+ and - strand labels to the left of the gene track. Default value is \code{strandLabels = TRUE}. Options = \list{TRUE, FALSE}}

\item{stroke}{numeric A numeric value indicating the stroke width for gene
body outlines. Default value is \code{stroke = 0.1}. Options = any numeric.}

\item{bg}{string Character value indicating background color.
Default value is \code{bg = NA}. Options = any string.}

\item{x}{numeric A numeric or unit object specifying genes plot x-location. Default value is \code{x = NULL}. Options = any numeric.}

\item{y}{numeric A numeric, unit object, or character containing a "b"
combined with a numeric value specifying genes plot y-location.
The character value will
place the genes plot y relative to the bottom of the most recently
plotted plot according to the units of the plotgardener page. Default value is \code{y = NULL}. Options = any numeric.}

\item{width}{numeric A numeric or unit object specifying genes plot width. Default value is \code{width = NULL}. Options \code{any numeric}.}

\item{height}{numeric A numeric or unit object specifying genes plot height. Default value is \code{height = NULL}. Options \code{any numeric}.}

\item{just}{Justification of genes plot relative to its (x, y) location.
If there are two values, the first value specifies horizontal
justification and the second value specifies vertical justification.
Possible string values are: \code{"left"}, \code{"right"},
\code{"centre"}, \code{"center"}, \code{"bottom"}, and \code{"top"}.
Default value is \code{just = c("left", "top")}. Options = 2 of \list{"left "top", "left "bottom", "left "center", "right "top", "right "bottom", "right "center"}.}

\item{default.units}{A string indicating the default units to use if
\code{x}, \code{y}, \code{width}, or \code{height} are only given
as numerics. Default value is \code{default.units = "inches"}. 
Option =  \text{"inches", "centimeters"}.}

\item{draw}{A logical value indicating whether graphics output
should be produced. Default value is \code{draw = TRUE}. Options = \list{TRUE, FALSE}.}

\item{params}{An optional \link[plotgardener]{pgParams} object
containing relevant function parameters. Default value is \code{params = NULL}.}
}
\value{
Returns a \code{genes} object containing
relevant genomic region, placement, and \link[grid]{grob} information.
}
\description{
Plot a gene track for a specified genomic region
}
\details{
A gene track can be placed on a page by providing
plot placement parameters:
\preformatted{
plotGenes(chrom, chromstart = NULL, chromend = NULL,
            x, y, width, height, just = c("left", "top"),
            default.units = "inches")
}
This function can be used to quickly plot an unnannotated gene track
by ignoring plot placement parameters:
\preformatted{
plotGenes(chrom, chromstart = NULL, chromend = NULL)
}

Genomic annotation information is acquired through
\link[GenomicFeatures]{TxDb} and \link[AnnotationDbi]{OrgDb-class}
packages, as determined
through the \code{assembly} parameter. To avoid overcrowding of gene name
labels, plotted gene labels are by default prioritized according to
citation counts.
}
\examples{
## Load hg19 genomic annotation packages
library("TxDb.Hsapiens.UCSC.hg19.knownGene")
library("org.Hs.eg.db")

## Set genomic coordinates
paramssmall <- pgParams(
    chrom = "chr8",
    chromstart = 1, chromend = 3000000,
    assembly = "hg19", width = 7
)
paramsbig <- pgParams(
    chrom = "chr8",
    chromstart = 1, chromend = 146364022,
    assembly = "hg19", width = 7
)
## Set colors
cols <- c("#41B6C4", "#225EA8")

## Create page
pageCreate(width = 7.5, height = 3.5, default.units = "inches")

## Plot genes big
genesPlot <- plotGenes(
    params = paramsbig, fill = cols,
    fontcolor = cols,
    x = 0.25, y = 0.25, height = 0.75,
    just = c("left", "top"),
    default.units = "inches"
)

## Annotate genome label
annoGenomeLabel(
    plot = genesPlot, x = 0.25, y = 1.0,
    scale = "Mb", just = c("left", "top")
)

## Plot genes small
genesPlot <- plotGenes(
    params = paramssmall,
    geneHighlights = data.frame(
        "gene" = c("DLGAP2"),
        "color" = c("#225EA8")
    ),
    geneBackground = "grey",
    x = 0.25, y = 2.25, height = 0.75,
    just = c("left", "top"), default.units = "inches"
)

## Annotate genome label
annoGenomeLabel(
    plot = genesPlot, x = 0.25, y = 3.0, scale = "Mb",
    just = c("left", "top")
)

## Hide page guides
pageGuideHide()
}
\seealso{
\link[plotgardener]{assembly},
\link[plotgardener]{genomes}, \link[plotgardener]{defaultPackages}
}
