% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCool.R
\name{readCoolBpResolutions}
\alias{readCoolBpResolutions}
\title{Read basepair resolutions from an .(m)cool file}
\usage{
readCoolBpResolutions(file)
}
\arguments{
\item{file}{A character value specifying the path to the .(m)cool file}
}
\value{
Vector of basepair resolutions
}
\description{
Read basepair resolutions from an .(m)cool file
}
\author{
Sarah Parker
}
