% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCool.R
\name{readCoolChroms}
\alias{readCoolChroms}
\title{Read chromosomes included in .(m)cool files}
\usage{
readCoolChroms(file, resolution = NULL)
}
\arguments{
\item{file}{A character value specifying the path to the .(m)cool file}

\item{resolution}{optional, specify which resolution(s) to read 
chromosomes from. Default is all resolutions in `file`.}
}
\value{
Data frame or list of data frames of chromosome names and lengths
}
\description{
Read chromosomes included in .(m)cool files
}
\author{
Sarah Parker, Nicole Kramer
}
