% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_addColorScale.R
\name{bb_addColorScale}
\alias{bb_addColorScale}
\title{adds color scale for heatmap plots}
\usage{
bb_addColorScale(
  plot,
  x,
  y,
  width,
  height,
  params = NULL,
  border = FALSE,
  orientation = "v",
  fontsize = 8,
  fontcolor = "dark grey",
  just = c("left", "top"),
  default.units = "inches",
  ...
)
}
\arguments{
\item{plot}{plot to add scale to}

\item{x}{A numeric vector or unit object specifying x-location}

\item{y}{A numeric vector or unit object specifying y-location}

\item{width}{A numeric vector or unit object specifying width}

\item{height}{A numeric vector or unit object specifying height}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{border}{option to add border around scale}

\item{orientation}{"v" (vertical) or "h" (horizontal) orientation}

\item{fontsize}{fontsize for text}

\item{fontcolor}{fontcolor for text}

\item{just}{justification of scale viewport}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numeric vectors}
}
\description{
adds color scale for heatmap plots
}
\author{
Nicole Kramer
}
