% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_addscale.R
\name{bb_addScale}
\alias{bb_addScale}
\title{adds color scale for heatmap plots}
\usage{
bb_addScale(plot, border = FALSE, height, width, x, y,
  orientation = "vertical", fontsize = 8, fontcolor = "grey",
  fontface = "plain", just = c("left", "top"), ...)
}
\arguments{
\item{plot}{plot to add scale to}

\item{border}{option to add border around scale}

\item{height}{A unit object specifying height}

\item{width}{A unit object specifying width}

\item{x}{A unit object specifying x-location}

\item{y}{A unit object specifying y-location}

\item{orientation}{"vertical" or "horizontal" orientation}

\item{fontsize}{fontsize for text}

\item{fontcolor}{fontcolor for text}

\item{fontface}{fontface for text}

\item{just}{justification of scale viewport}
}
\description{
adds color scale for heatmap plots
}
\author{
Nicole Kramer
}
