% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_addText.R
\name{bb_addText}
\alias{bb_addText}
\title{wrapper to draw a textGrob based on bb_makePage coordinates and units}
\usage{
bb_addText(label, x, y, just = "center", rotation = 0,
  units = "inches", fontcolor = "black", transparency = 1,
  fontsize = 12, cex = 1, fontfamily = "", fontface = "plain")
}
\arguments{
\item{label}{character or expression}

\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{just}{justification of text relative to its (x, y) location}

\item{rotation}{angle to rotate text}

\item{units}{units of x and y coordinates}

\item{fontcolor}{fontcolor}

\item{transparency}{degree of text transparency}

\item{fontsize}{the size of text (in points)}

\item{cex}{multiplier applied to fontsize}

\item{fontfamily}{the font family}

\item{fontface}{the fontface (bold, italic, ...)}
}
\description{
wrapper to draw a textGrob based on bb_makePage coordinates and units
}
