% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_defaultPackages.R
\name{bb_defaultPackages}
\alias{bb_defaultPackages}
\title{Displays the default data packages associated with a genome build}
\usage{
bb_defaultPackages(genome)
}
\arguments{
\item{genome}{string indicating the genome assembly}
}
\description{
Displays the default data packages associated with a genome build
}
