% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_hGuide.R
\name{bb_hGuide}
\alias{bb_hGuide}
\title{draws a horizontal guideline at a specified y-coordinate}
\usage{
bb_hGuide(y, units = "inches", col = "grey55")
}
\arguments{
\item{y}{y-coordinate to draw guide}

\item{units}{units of y-coordinate}

\item{col}{color of guideline}
}
\description{
draws a horizontal guideline at a specified y-coordinate
}
