% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_hguide.R
\name{bb_hGuide}
\alias{bb_hGuide}
\title{draws a horizontal guideline at a specified y-coordinate}
\usage{
bb_hGuide(y, col = "grey55", default.units = "inches")
}
\arguments{
\item{y}{A numeric or unit object specifying y-coordinate of guide}

\item{col}{color of guideline}

\item{default.units}{A string indicating the default units to use if y is only given as numeric vectors}
}
\description{
draws a horizontal guideline at a specified y-coordinate
}
