% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_hic.R
\name{bb_hic}
\alias{bb_hic}
\title{plots HiC interaction matrix}
\usage{
bb_hic(hic, chrom = "chr8", chromstart = 133600000,
  chromend = 134800000, altchrom = NULL, altchromstart = NULL,
  altchromend = NULL, althalf = "bottom", resolution = 10000,
  zrange = NULL, height = 3, width = 3, x = 1.75, y = 3,
  units = "inches", palette = colorRampPalette(c("white", "dark red")),
  norm = "KR", half = "both", raster = TRUE, addlegend = FALSE,
  legendlocation = "right", legendoffset = 0.25, ...)
}
\arguments{
\item{hic}{path to .hic file or 3 column dataframe of counts}

\item{chrom}{chromosome of region to be plotted}

\item{chromstart}{chromosome start of region to be plotted}

\item{chromend}{chromosome end of region to be plotted}

\item{resolution}{the width in bp of each pixel}

\item{zrange}{the range of interaction scores to plot, where extreme values will be set to the max or min}

\item{height}{height of plot in inches}

\item{width}{width of plot in inches}

\item{x}{x-coordinate of where to place plot relative to top left of plot}

\item{y}{y-coordinate of where to place plot relative to top left of plot}

\item{units}{units of height, width, and x and y location of the plot}

\item{palette}{ColorRamp palette to use for representing interaction scores}

\item{norm}{if giving .hic file, hic data normalization; options are "NONE", "VC", "VC_SQRT", and "KR"}

\item{half}{what sides of square plot; options are "both", top", or "bottom"}

\item{raster}{allows for rasterization of plot, which results in quicker plotting}

\item{addlegend}{add legend representing scores for color range}

\item{legendlocation}{if addlegend == TRUE, where relative to plot to place legend; options are "right", "left", top", or "bottom"}

\item{legendoffset}{if addlegend == TRUE, how much to offset the legend relative to the plot in inches}
}
\value{
Function will plot a HiC interaction matrix and return a list of the zrange min, zrange max, and sequential color palette
}
\description{
plots HiC interaction matrix
}
\details{
macOS Preview anti-aliases images and will make rasterized plot appear blurry.
}
\examples{
data(bb_hic_data)
# Full Rasterized Plot
bb_makepage()
bb_hic_plot <- bb_hic(bb_hic_data)
# Full Rasterized Plot with Legend
bb_makepage()
bb_hic_plot <- bb_hic(bb_hic_data, addlegend = TRUE)
# Half Rasterized Plot
bb_makepage()()
bb_hic_plot <- bb_hic(bb_hic_data, half = "top")
# Non-rasterized Plot
bb_makepage()
bb_hic_plot <- bb_hic(bb_hic_data, raster = FALSE)

}
\author{
Nicole Kramer
}
