% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_labelGenome.R
\name{bb_labelGenome}
\alias{bb_labelGenome}
\title{Adds genome coordinates to the axis of a plot}
\usage{
bb_labelGenome(plot = NULL, chrom, chromstart, chromend, width, x, y,
  just = c("left", "top"), rotation = 0, units = "inches",
  scale = "bp", fontsize = 10, fontcolor = "black", lineAbove = T,
  linecolor = "black", lwd = 1)
}
\arguments{
\item{plot}{plot to annotate (saved as output from BentoBox plotting function)}

\item{chrom}{if not inputting plot, chromosome of region to plot}

\item{chromstart}{if not inputting plot, chromstart of region to plot}

\item{chromend}{if not inputting plot, chromend of region to plot}

\item{width}{if not inputting plot, width of label}

\item{x}{if not inputting plot, x-coordinate of label}

\item{y}{if not inputting plot, y-coordinate of label}

\item{just}{justification}

\item{units}{units of inputted x and y}

\item{scale}{scale of the plot; options are "bp", "Kb", or "Mb"}

\item{fontsize}{fontsize for labels (in points)}

\item{fontcolor}{color for all text/lines}

\item{lineAbove}{TRUE/FALSE indicating if line should be above or below text}

\item{linecolor}{linecolor}

\item{lwd}{linewidth}

\item{side}{side of the plot to annotate, options are "bottom", "top", "right", or "left"}
}
\description{
Adds genome coordinates to the axis of a plot
}
