% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_labelGenome.R
\name{bb_labelGenome}
\alias{bb_labelGenome}
\title{Adds genome coordinates to the axis of a plot}
\usage{
bb_labelGenome(plot, x, y, just = c("left", "top"), rotation = 0,
  scale = "bp", fontsize = 10, fontcolor = "black", lineAbove = T,
  linecolor = "black", lwd = 1)
}
\arguments{
\item{plot}{plot to annotate}

\item{x}{A unit object specifying x-location}

\item{y}{A unit object specifying y-location}

\item{just}{justification}

\item{scale}{scale of the plot; options are "bp", "Kb", or "Mb"}

\item{fontsize}{fontsize for labels (in points)}

\item{fontcolor}{color for all text/lines}

\item{lineAbove}{TRUE/FALSE indicating if line should be above or below text}

\item{linecolor}{linecolor}

\item{lwd}{linewidth}

\item{units}{units of inputted x and y}

\item{side}{side of the plot to annotate, options are "bottom", "top", "right", or "left"}
}
\description{
Adds genome coordinates to the axis of a plot
}
