% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_labelGenome.R
\name{bb_labelGenome}
\alias{bb_labelGenome}
\title{Adds genome coordinates to the axis of a plot}
\usage{
bb_labelGenome(
  plot,
  x,
  y,
  just = c("left", "top"),
  rotation = 0,
  scale = "bp",
  fontsize = 10,
  fontcolor = "black",
  linecolor = "black",
  lwd = 1,
  fontfamily = "",
  commas = TRUE,
  ticks = NULL,
  tcl = 0.5
)
}
\arguments{
\item{plot}{plot to annotate}

\item{x}{A unit object specifying x-location}

\item{y}{A unit object specifying y-location}

\item{just}{justification}

\item{rotation}{angle of rotation of label}

\item{scale}{scale of the plot; options are "bp", "Kb", or "Mb"}

\item{fontsize}{fontsize for labels (in points)}

\item{fontcolor}{color for all text/lines}

\item{linecolor}{linecolor}

\item{lwd}{linewidth}

\item{fontfamily}{fontfamily for text}

\item{commas}{A logical value indicating whether to include commas in start and stop labels}

\item{ticks}{Specified locations of ticks}

\item{tcl}{Length of tickmark as fraction of text height}
}
\description{
Adds genome coordinates to the axis of a plot
}
