% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_labelgenome.R
\name{bb_labelGenome}
\alias{bb_labelGenome}
\title{Adds genome coordinates to the axis of a BentoBox plot}
\usage{
bb_labelGenome(
  plot,
  x,
  y,
  params = NULL,
  just = c("left", "top"),
  scale = "bp",
  fontsize = 10,
  fontcolor = "black",
  linecolor = "black",
  commas = TRUE,
  sequence = TRUE,
  boxWidth = 0.5,
  ticks = NULL,
  tcl = 0.5,
  default.units = "inches",
  ...
)
}
\arguments{
\item{plot}{BentoBox plot to annotate.}

\item{x}{A numeric or unit object specifying x-location.}

\item{y}{A numeric or unit object specifying y-location.}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{just}{The justification of the label relative to its (x, y) location.If there are two values, the first specifies horizontal justification and
the second value specifies vertical justification. Possible string values are: "left", "right", "centre", "center", "bottom", and "top".}

\item{scale}{scale of the genome of the label. Options are "bp", "Kb", or "Mb"}

\item{fontsize}{Fontsize for labels (in points).}

\item{fontcolor}{Color for all text and lines (with the exception of sequence information).}

\item{linecolor}{Axis linecolor.}

\item{commas}{A logical value indicating whether to include commas in start and stop labels.}

\item{sequence}{A logical value indicating whether to include sequence information above the label axis (only at appropriate resolutions).}

\item{ticks}{A numeric vector of x-value locations for tick marks.}

\item{tcl}{Length of tickmark as fraction of text height.}

\item{default.units}{A string indicating the default units to use if x or y are only given as numerics.}

\item{lwd}{Axis linewidth.}

\item{fontfamily}{Text fontfamily.}
}
\description{
Adds genome coordinates to the axis of a BentoBox plot
}
