% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_legend.R
\name{bb_legend}
\alias{bb_legend}
\title{adds legend to HiC interaction matrix}
\usage{
bb_legend(color_vector, min_label, max_label, height, width, x, y,
  pageheight, orientation = "vertical", fontsize = 8,
  fontcolor = "grey", ...)
}
\arguments{
\item{color_vector}{sequential list of colors used to represent interaction frequency; returned from bb_hic function}

\item{min_label}{label for minimum value, i.e first value of zrange}

\item{max_label}{label for maximum value, i.e. last value of zrange}

\item{height}{legend height, in inches}

\item{width}{legend width, in inches}

\item{x}{legend x-coordinate}

\item{y}{legend y-coordinate}

\item{pageheight}{height of page on which plot will be placed}

\item{orientation}{"vertical" or "horizontal" orientation}

\item{fontsize}{fontsize for text}

\item{fontcolor}{fontcolor for test}
}
\description{
adds legend to HiC interaction matrix
}
\author{
Nicole Kramer
}
