% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_makepage.R
\name{bb_makepage}
\alias{bb_makepage}
\title{generates new page for plotting}
\usage{
bb_makepage(width = 8.5, height = 11, units = "inches",
  showOutline = FALSE)
}
\arguments{
\item{width}{width of page}

\item{height}{height of page}

\item{units}{units for height and width; options are same as grid options: "inches", "cm", "mm", and "npc"}

\item{showOutline}{if TRUE, will show limits of defined page}
}
\description{
generates new page for plotting
}
