% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pageGuideHide.R
\name{bb_pageGuideHide}
\alias{bb_pageGuideHide}
\title{Remove guides from a BentoBox page}
\usage{
bb_pageGuideHide()
}
\value{
None.
}
\description{
Remove guides from a BentoBox page
}
\examples{
## Make a BentoBox page
bb_pageCreate(width = 7, height = 4, default.units = "inches")

## Hide page guides
bb_pageGuideHide()

}
