% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pageGuideShow.R
\name{bb_pageGuideShow}
\alias{bb_pageGuideShow}
\title{Reshow guides drawn with \code{bb_pageCreate}, \code{bb_pageGuideHorizontal}, and \code{bb_pageGuideVertical}}
\usage{
bb_pageGuideShow()
}
\description{
Reshow guides drawn with \code{bb_pageCreate}, \code{bb_pageGuideHorizontal}, and \code{bb_pageGuideVertical}
}
\seealso{
\link[BentoBox]{bb_pageCreate}, \link[BentoBox]{bb_pageGuideHorizontal}, \link[BentoBox]{bb_pageGuideVertical}
}
