% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pagePlotPlace.R
\name{bb_pagePlotPlace}
\alias{bb_pagePlotPlace}
\title{Place a BentoBox plot that has been previously created but not drawn}
\usage{
bb_pagePlotPlace(
  plot,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  params = NULL
)
}
\arguments{
\item{plot}{BentoBox plot object to be placed, defined by the output of a BentoBox plotting function.}

\item{x}{A numeric or unit object specifying plot x-location.}

\item{y}{A numeric, unit object, or character containing a "b" combined with a numeric value specifying plot y-location. The character value will
place the plot y relative to the bottom of the most recently plotted BentoBox plot according to the units of the BentoBox page.}

\item{width}{A numeric or unit object specifying plot width.}

\item{height}{A numeric or unit object specifying plot height.}

\item{just}{Justification of plot relative to its (x, y) location. If there are two values, the first value specifies horizontal justification and the second value specifies vertical justification.
Possible string values are: \code{"left"}, \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"}, and \code{"top"}. Default value is \code{just = c("left", "top")}.}

\item{default.units}{A string indicating the default units to use if \code{x}, \code{y}, \code{width}, or \code{height} are only given as numerics. Default value is \code{default.units = "inches"}.}

\item{draw}{A logical value indicating whether graphics output should be produced. Default value is \code{draw = TRUE}.}

\item{params}{An optional \link[BentoBox]{bb_params} object containing relevant function parameters.}
}
\value{
Function will update dimensions of an input plot and return an updated BentoBox plot object.
}
\description{
Place a BentoBox plot that has been previously created but not drawn
}
\examples{
## Load Hi-C data
data("bb_hicData")

## Create, but do not plot, square Hi-C plot
hicPlot <- bb_plotHicSquare(data = bb_hicData, resolution = 10000, zrange = c(0, 70),
                            chrom = "chr21", chromstart = 28000000, chromend = 30300000,
                            draw = FALSE)

## Create BentoBox page
bb_pageCreate(width = 4, height = 3.5, default.units = "inches")

## Place Hi-C plot on BentoBox page
bb_pagePlotPlace(plot = hicPlot,
                 x = 0.5, y = 0.5, width = 2.5, height = 2.5,
                 just = c("left", "top"), default.units = "inches", draw = TRUE)

}
