% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_placePlot.R
\name{bb_placePlot}
\alias{bb_placePlot}
\title{places a plot that has been previously created but not drawn}
\usage{
bb_placePlot(
  plot,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  draw = T
)
}
\arguments{
\item{plot}{plot to be placed}

\item{x}{A unit object specifying x-location.}

\item{y}{A unit object specifying y-location.}

\item{width}{A unit object specifying width.}

\item{height}{A unit object specifying height.}

\item{just}{A string or numeric vector specifying the justification of the plot relative to its (x, y) location}
}
\value{
Function will plot a HiC interaction matrix and return a bb_hicPlot object
}
\description{
places a plot that has been previously created but not drawn
}
