% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotApa.R
\name{bb_plotApa}
\alias{bb_plotApa}
\title{plots an APA plot}
\usage{
bb_plotApa(
  apa,
  params = NULL,
  loopNumber = 1,
  palette = colorRampPalette(c("white", "dark red")),
  scaleStrength = TRUE,
  zrange = NULL,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("center"),
  default.units = "inches",
  draw = TRUE
)
}
\arguments{
\item{apa}{path to APA txt file or an APA matrix}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{loopNumber}{number of DNA loops}

\item{palette}{ColorRamp palette to use for representing interaction scores}

\item{scaleStrength}{A logical value indicating divide aggregate values by number of loops to get average loop strength}

\item{zrange}{the range of interaction scores to plot, where extreme values will be set to the max or min}

\item{x}{A numeric or unit object specifying x-location}

\item{y}{A numeric or unit object specifying y-location}

\item{width}{A numeric or unit object specifying width}

\item{height}{A numeric or unit object specifying height}

\item{just}{justification of x and y-coordinates}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numeric vectors}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\description{
plots an APA plot
}
