% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotApa.R
\name{bb_plotApa}
\alias{bb_plotApa}
\title{plots an Apa plot}
\usage{
bb_plotApa(
  apa,
  loopNumber = 1,
  palette = colorRampPalette(c("white", "dark red")),
  zrange = NULL,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("center"),
  draw = TRUE
)
}
\arguments{
\item{apa}{path to APA txt file or an APA matrix}

\item{loopNumber}{number of DNA loops}

\item{palette}{ColorRamp palette to use for representing interaction scores}

\item{x}{A unit object specifying x-location}

\item{y}{A unit object specifying y-location}

\item{width}{A unit object specifying width}

\item{height}{A unit object specifying height}

\item{just}{justification of x and y-coordinates}

\item{draw}{A logical value indicating whether graphics output should be produced}

\item{zrangethe}{range of interaction scores to plot, where extreme values will be set to the max or min}
}
\description{
plots an Apa plot
}
