% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotBedpe.R
\name{bb_plotBedpe}
\alias{bb_plotBedpe}
\title{plots paired-end data for a single chromosome}
\usage{
bb_plotBedpe(
  bedpe,
  chrom,
  params = NULL,
  chromstart = NULL,
  chromend = NULL,
  fillcolor = "black",
  colorby = NULL,
  colorbyrange = NULL,
  linecolor = NA,
  assembly = "hg19",
  boxHeight = unit(2, "mm"),
  spaceHeight = 0.3,
  spaceWidth = 0.02,
  baseline = FALSE,
  bg = NA,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  ...
)
}
\arguments{
\item{bedpe}{bed paired end data to be plotted}

\item{chrom}{chromsome of region to be plotted}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{chromstart}{start position}

\item{chromend}{end position}

\item{fillcolor}{single value, vector, or palette specifying colors of bedpe elements}

\item{colorby}{name of column in bedpe data to scale fillcolors by}

\item{colorbyrange}{the range of values to apply a colorby palette scale to, if colorby values are numeric}

\item{linecolor}{border color}

\item{assembly}{default genome assembly as a string or a bb_assembly object}

\item{boxHeight}{A numeric or unit object specifying height of boxes at either end of bedpe element}

\item{spaceHeight}{height of space between boxes of different bedpe elements}

\item{spaceWidth}{width of spacing between bedpe elements, as a fraction of the plot's genomic range}

\item{baseline}{logical indicating whether to include baseline along the x-axis}

\item{bg}{background color}

\item{x}{A numeric vector or unit object specifying x-location}

\item{y}{A numeric vector or unit object specifying y-location}

\item{width}{A numeric vector or unit object specifying width}

\item{height}{A numeric vector or unit object specifying height}

\item{just}{string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numeric vectors}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\description{
plots paired-end data for a single chromosome
}
