% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotBedpe.R
\name{bb_plotBedpe}
\alias{bb_plotBedpe}
\title{plots paired-end data for a single chromosome}
\usage{
bb_plotBedpe(
  bedpe,
  chrom,
  chromstart,
  chromend,
  fillcolor = "black",
  colorby = NULL,
  colorbycol = NULL,
  colorbyrange = NULL,
  linecolor = NA,
  boxHeight = unit(0.025, "inches"),
  spaceHeight = unit(0.025, "inches"),
  limitDots = TRUE,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  ...
)
}
\arguments{
\item{bedpe}{bed paired end data to be plotted}

\item{chrom}{chromsome of region to be plotted}

\item{chromstart}{start position}

\item{chromend}{end position}

\item{fillcolor}{single value or vector specifying colors of bedpe elements}

\item{colorby}{vector to scale colors by}

\item{colorbycol}{palette to apply color scale to (only valid when colorby is not NULL)}

\item{colorbyrange}{the range of values to apply the color scale to}

\item{linecolor}{border color}

\item{boxHeight}{height of boxes at either end of bedpe element}

\item{spaceHeight}{height of space between boxes of different bedpe elements}

\item{limitDots}{logical value indicating whether to plot "..." to indicate additional, unplotted bedpe elements}

\item{x}{A numeric vector or unit object specifying x-location}

\item{y}{A numeric vector or unit object specifying y-location}

\item{width}{A numeric vector or unit object specifying width}

\item{height}{A numeric vector or unit object specifying height}

\item{just}{string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numeric vectors}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\description{
plots paired-end data for a single chromosome
}
