% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotBedpe.R
\name{bb_plotBedpe}
\alias{bb_plotBedpe}
\title{Plot paired-end BEDPE data for a single chromosome}
\usage{
bb_plotBedpe(
  data,
  chrom,
  chromstart = NULL,
  chromend = NULL,
  assembly = "hg19",
  fill = "black",
  colorby = NULL,
  linecolor = NA,
  bg = NA,
  boxHeight = unit(2, "mm"),
  spaceWidth = 0.02,
  spaceHeight = 0.3,
  baseline = FALSE,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  params = NULL,
  ...
)
}
\arguments{
\item{data}{A string specifying the BEDPE file path or a dataframe in BEDPE format specifying data to be plotted.}

\item{chrom}{Chromosome of region to be plotted, as a string.}

\item{chromstart}{Integer start position on chromosome to be plotted.}

\item{chromend}{Integer end position on chromosome to be plotted.}

\item{assembly}{Default genome assembly as a string or a \link[BentoBox]{bb_assembly} object. Default value is \code{assembly = "hg19"}.}

\item{fill}{Character value(s) as a single value, vector, or palette specifying fill colors of BEDPE elements. Default value is \code{fill = "black"}.}

\item{colorby}{A "\link[BentoBox]{colorby}" object specifying information for scaling colors in data.}

\item{linecolor}{A character value specifying the color of the lines outlining BEDPE elements. Default value is \code{linecolor = NA}.}

\item{bg}{Character value indicating background color. Default value is \code{bg = NA}.}

\item{boxHeight}{A numeric or unit object specifying height of boxes at either end of BEDPE elements. Default value is \code{boxHeight = unit(2, "mm")}.}

\item{spaceWidth}{A numeric specifying the width of spacing between BEDPE elements, as a fraction of the plot's genomic range. Default value is \code{spaceWidth = 0.02}.}

\item{spaceHeight}{A numeric specifying the height of space between boxes of BEDPE elements on different rows. Default value is \code{spaceHeight = 0.3}.}

\item{baseline}{Logical value indicating whether to include a baseline along the x-axis. Default value is \code{baseline = FALSE}.}

\item{x}{A numeric or unit object specifying BEDPE plot x-location.}

\item{y}{A numeric or unit object specifying BEDPE plot y-location.}

\item{width}{A numeric or unit object specifying BEDPE plot width.}

\item{height}{A numeric or unit object specifying BEDPE plot height.}

\item{just}{Justification of BEDPE plot relative to its (x, y) location. If there are two values, the first value specifies horizontal justification and the second value specifies vertical justification.
Possible string values are: \code{"left"}, \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"}, and \code{"top"}. Default value is \code{just = c("left", "top")}.}

\item{default.units}{A string indicating the default units to use if \code{x}, \code{y}, \code{width}, or \code{height} are only given as numerics. Default value is \code{default.units = "inches"}.}

\item{draw}{A logical value indicating whether graphics output should be produced.}

\item{params}{An optional \link[BentoBox]{bb_params} object containing relevant function parameters.}

\item{...}{Additional grid graphical parameters. See \link[grid]{gpar}.}
}
\value{
Returns a \code{bb_bedpe} object containing relevant genomic region, placement, and \link[grid]{grob} information.
}
\description{
Plot paired-end BEDPE data for a single chromosome
}
\details{
#' A BEDPE plot can be placed on a BentoBox coordinate page by providing plot placement parameters:
\preformatted{
bb_plotBedpe(data, chrom,
             chromstart = NULL, chromend = NULL,
             x, y, width, height, just = c("left", "top"),
             default.units = "inches")
}
This function can also be used to quickly plot an unannotated BEDPE plot by ignoring plot placement parameters:
\preformatted{
bb_plotBedpe(data, chrom,
             chromstart = NULL, chromend = NULL)
}
}
\examples{
## Load BEDPE data
data("bb_bedpeData")

## Set the coordinates
params = bb_params(chrom = "chr21",chromstart = 27900000, chromend = 30700000,
                   width = 7)

## Create a page
bb_pageCreate(width = 7.5, height = 2.1, default.units = "inches")

## Add a length column
bb_bedpeData$length = (bb_bedpeData$start2 - bb_bedpeData$start1) / 1000

## Plot the data
bedpePlot <- bb_plotBedpe(data = bb_bedpeData, params = params,
                          fill = colorRampPalette(c("dodgerblue2", "firebrick2")),
                          colorby = colorby("length"),
                          lwd = 2, spaceHeight = .7,
                          x = 0.25, y = 0.25, height = 1.5,
                          just = c("left", "top"), default.units = "inches")

## Annotate genome label
bb_annoGenomeLabel(plot = bedpePlot, x = 0.25, y = 1.78, scale = "Mb")

## Add heatmap legend
bb_annoHeatmapLegend(plot = bedpePlot, fontcolor = "black", x = 7.0, y = 0.25,
                     width = 0.10, height = 1,fontsize=10)

## Add heatmap legend label
bb_plotText(label = "Kb", rot = 90, x = 6.9, y = 0.75,
             just = c("center","center"), fontsize = 10)

## Hide page guides
bb_pageGuideHide()

}
