% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotBedpeArches.R
\name{bb_plotBedpeArches}
\alias{bb_plotBedpeArches}
\title{Plot paired-end BEDPE data in an arch style}
\usage{
bb_plotBedpeArches(bedpeData, chrom)
bb_plotBedpeArches(bedpeData, chrom, x, y, width,
                   height, just = c("left", "top"),
                   default.units = "inches")
}
\arguments{
\item{bedpeData}{A string specifying the BEDPE file path or a dataframe in BEDPE format specifying data to be plotted.}

\item{chrom}{Chromosome of region to be plotted, as a string.}

\item{chromstart}{Integer start position on chromosome to be plotted.}

\item{chromend}{Integer end position on chromosome to be plotted.}

\item{assembly}{Default genome assembly as a string or a \link[BentoBox]{bb_assembly} object. Default value is \code{assembly = "hg19"}.}

\item{style}{Character value describing the style of arches. Default value is \code{style = "2D"}. Options are:
\itemize{
\item{\code{"2D"}: }{Arches will be drawn in a 2-dimensional style.}
\item{\code{"3D"}: }{Arches will be drawn in a 3-dimensional style.}
}}

\item{position}{Character value indicating the position of arches relative to the x-axis. Default value is \code{position = "top"}. Options are:
\itemize{
\item{\code{"top"}: }{Arches will be drawn above the x-axis.}
\item{\code{"bottom"}: }{Arches will be drawn reflected over and below the x-axis.}
}}

\item{curvature}{Numeric indicating the number of points along the arch curvature. Default value is \code{curvature = 5}.}

\item{archHeight}{Single numeric value or numeric vector specifying the arch heights. When NULL, all arches will be the same height, filling up the given plot area}

\item{fill}{Character value(s) as a single value, vector, or palette specifying fill colors of arches. Default value is \code{fill = "grey"}.}

\item{colorby}{A "\link[BentoBox]{colorby}" object specifying information for scaling colors in data.}

\item{linecolor}{A character value specifying the color of the lines outlining arches. Default value is \code{linecolor = "grey"}.}

\item{alpha}{Numeric value specifying transparency. Default value is \code{alpha = 0.4}.}

\item{bg}{Character value indicating background color. Default value is \code{bg = NA}.}

\item{clip}{A logical value indicating whether to clip any arches that get cutoff in the given genomic region. Default value is \code{clip = TRUE}.}

\item{baseline}{Logical value indicating whether to include a baseline along the x-axis. Default value is \code{baseline = FALSE}.}

\item{x}{A numeric or unit object specifying BEDPE arches plot x-location.}

\item{y}{A numeric or unit object specifying BEDPE arches plot y-location.}

\item{width}{A numeric or unit object specifying BEDPE arches plot width.}

\item{height}{A numeric or unit object specifying BEDPE arches plot height.}

\item{just}{Justification of BEDPE arches plot relative to its (x, y) location. If there are two values, the first value specifies horizontal justification and the second value specifies vertical justification.
Possible string values are: \code{"left"}, \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"}, and \code{"top"}. Default value is \code{just = c("left", "top")}.}

\item{default.units}{A string indicating the default units to use if \code{x}, \code{y}, \code{width}, or \code{height} are only given as numerics. Default value is \code{default.units = "inches"}.}

\item{draw}{A logical value indicating whether graphics output should be produced. Default value is \code{draw = TRUE}.}

\item{params}{An optional \link[BentoBox]{bb_assembly} object containing relevant function parameters.}

\item{...}{Additional grid graphical parameters. See \link[grid]{gpar}.}
}
\value{
Returns a \code{bb_arches} object containing relevant genomic region, placement, and \link[grid]{grob} information.
}
\description{
Plot paired-end BEDPE data in an arch style
}
\examples{
## Load BEDPE data
data("bb_bedpeData")

## Plot BEDPE arches plot filling up entire graphic device
bedpeArchesPlot <- bb_plotBedpeArches(bedpeData = bb_bedpeData, chrom = "chr21", chromstart = 28000000, chromend = 30300000, fill = "black")

## Plot and place BEDPE arches plot on a BentoBox page
bb_pageCreate(width = 5, height = 2, default.units = "inches", xgrid = 0, ygrid = 0)
bedpeArchesPlot <- bb_plotBedpeArches(bedpeData = bb_bedpeData, chrom = "chr21", chromstart = 28000000, chromend = 30300000, fill = "black",
                                      x = 0, y = 0.25, width = 5, height = 1.5, just = c("left", "top"), default.units = "inches")

}
