% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotCircle.R
\name{bb_plotCircle}
\alias{bb_plotCircle}
\title{wrapper to draw a circleGrob based on BentoBox page coordinates and units}
\usage{
bb_plotCircle(
  x,
  y,
  r,
  params = NULL,
  linecolor = "black",
  fill = NA,
  lwd = 1,
  lty = 1,
  alpha = 1,
  default.units = "inches",
  ...
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-location of the center of the circle}

\item{y}{A numeric vector or unit object specifying y-location of the center of the circle}

\item{r}{A numeric vector or unit object specifying radii}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{linecolor}{line color}

\item{fill}{fill color}

\item{lwd}{line width}

\item{lty}{line type}

\item{alpha}{color transparency}

\item{default.units}{A string indicating the default units to use if x or y are only given as numeric vectors}
}
\description{
wrapper to draw a circleGrob based on BentoBox page coordinates and units
}
