% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotgenes.R
\name{bb_plotGenes}
\alias{bb_plotGenes}
\title{plots a track of genes}
\usage{
bb_plotGenes(
  gtf,
  chrom = "chr8",
  chromstart = 133600000,
  chromend = 134800000,
  fontcolors = c("#ff3434", "#2929ff"),
  strandcolors = c("#ff7e7e", "#8a8aff"),
  width = unit(3, "inches"),
  height = unit(1, "inches"),
  x = unit(1, "inches"),
  y = unit(1, "inches"),
  just = c("left", "top"),
  fontsize = 12,
  exclude = ""
)
}
\arguments{
\item{gtf}{gtf file or dataframe}

\item{chrom}{chromsome of gene region as a string}

\item{chromstart}{chromstart of gene region}

\item{chromend}{chromend of gene region}

\item{fontcolors}{a vector of 2 indicating the font colors for the plus and minus strands}

\item{strandcolors}{a vector of 2 indicating the strang color for the plus and minus strands}

\item{width}{width of plot}

\item{height}{height of plot}

\item{x}{x-coordinate of plot}

\item{y}{y-coordinate of plot}

\item{fontsize}{fontsize of gene names}

\item{exclude}{names of genes to exclude from gene labeling}
}
\description{
plots a track of genes
}
