% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotGenes.R
\name{bb_plotGenes}
\alias{bb_plotGenes}
\title{plots a gene track for a specified region}
\usage{
bb_plotGenes(
  chrom,
  params = NULL,
  assembly = "hg19",
  chromstart = NULL,
  chromend = NULL,
  fontcolors = c("#2929ff", "#ff3434"),
  strandcolors = c("#8a8aff", "#ff7e7e"),
  geneOrder = NULL,
  geneHighlights = NULL,
  geneBackground = "grey",
  stroke = 0.1,
  fontsize = 8,
  strandLabels = TRUE,
  x = NULL,
  y = NULL,
  width = NULL,
  height = unit(0.6, "inches"),
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE
)
}
\arguments{
\item{chrom}{chromsome of region to be plotted}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{assembly}{desired genome assembly}

\item{chromstart}{start position}

\item{chromend}{end position}

\item{fontcolors}{a vector indicating the font colors for the plus strand and minus strand gene labels}

\item{strandcolors}{a vector indicating the strand colors for the plus strand and minus strand}

\item{geneOrder}{an ordered vector of gene names to prioritize labeling}

\item{geneHighlights}{a two-column dataframe with gene names to highlight and their corresponding highlight color}

\item{geneBackground}{if geneHighlights is given, background color for genes that are not highlighted}

\item{stroke}{numerical value indicating the stroke width for gene body outlines}

\item{fontsize}{the size of gene label text (in points)}

\item{strandLabels}{A logical value indicating whether to include +/- strand labels}

\item{x}{A numeric or unit object specifying x-location}

\item{y}{A numeric or unit object specifying y-location}

\item{width}{A numeric or unit object specifying width}

\item{height}{A numeric or unit object specifying height}

\item{just}{string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numerics}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\value{
Function will plot a gene track and return a bb_genes object
}
\description{
plots a gene track for a specified region
}
