% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotGenes.R
\name{bb_plotGenes}
\alias{bb_plotGenes}
\title{plots a track of genes}
\usage{
bb_plotGenes(gtf, chrom = "chr8", chromstart = 133600000,
  chromend = 134800000, fontcolors = c("#ff3434", "#2929ff"),
  strandcolors = c("#ff7e7e", "#8a8aff"), width = 3, height = 1,
  x = 1, y = 1, just = c("left", "top"), units = "inches",
  fontsize = 12, exclude = "")
}
\arguments{
\item{gtf}{gtf file or dataframe}

\item{chrom}{chromsome of gene region as a string}

\item{chromstart}{chromstart of gene region}

\item{chromend}{chromend of gene region}

\item{fontcolors}{a vector of 2 indicating the font colors for the plus and minus strands}

\item{strandcolors}{a vector of 2 indicating the strang color for the plus and minus strands}

\item{width}{width of plot}

\item{height}{height of plot}

\item{x}{x-coordinate of plot}

\item{y}{y-coordinate of plot}

\item{just}{justification of plot}

\item{units}{units of width, height, x, and y-coordinates}

\item{fontsize}{fontsize of gene names}

\item{exclude}{names of genes to exclude from gene labeling}
}
\description{
plots a track of genes
}
