% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotGenomeLabel.R
\name{bb_plotGenomeLabel}
\alias{bb_plotGenomeLabel}
\title{Plots genomic coordinates for the axis of a BentoBox plot}
\usage{
bb_plotGenomeLabel(
  x,
  y,
  plot = NULL,
  width = NULL,
  chrom = NULL,
  chromstart = NULL,
  chromend = NULL,
  params = NULL,
  just = c("left", "top"),
  scale = "bp",
  assembly = "hg19",
  fontsize = 10,
  fontcolor = "black",
  linecolor = "black",
  commas = TRUE,
  sequence = TRUE,
  boxWidth = 0.5,
  ticks = NULL,
  tcl = 0.5,
  default.units = "inches",
  ...
)
}
\arguments{
\item{x}{A numeric or unit object specifying x-location.}

\item{y}{A numeric or unit object specifying y-location.}

\item{plot}{BentoBox plot to add genome label}

\item{width}{if not specifying plot, width of genome label, as a numeric or unit object}

\item{chrom}{if not specifying plot, chromosome of genome label}

\item{chromstart}{if not specifying plot, chromstart of genome label}

\item{chromend}{if not specifying plot, chromend of genome label}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{just}{The justification of the label relative to its (x, y) location.If there are two values, the first specifies horizontal justification and
the second value specifies vertical justification. Possible string values are: "left", "right", "centre", "center", "bottom", and "top".}

\item{scale}{scale of the genome of the label. Options are "bp", "Kb", or "Mb"}

\item{assembly}{default genome assembly as a string or a bb_assembly object (will be inherited from plot object if provided)}

\item{fontsize}{Fontsize for labels (in points).}

\item{fontcolor}{Color for all text and lines (with the exception of sequence information).}

\item{linecolor}{Axis linecolor.}

\item{commas}{A logical value indicating whether to include commas in start and stop labels.}

\item{sequence}{A logical value indicating whether to include sequence information above the label axis (only at appropriate resolutions).}

\item{ticks}{A numeric vector of x-value locations for tick marks.}

\item{tcl}{Length of tickmark as fraction of text height.}

\item{default.units}{A string indicating the default units to use if x or y are only given as numerics.}

\item{lwd}{Axis linewidth.}

\item{fontfamily}{Text fontfamily.}
}
\description{
Plots genomic coordinates for the axis of a BentoBox plot
}
