% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotGG.R
\name{bb_plotGG}
\alias{bb_plotGG}
\title{adds a plot created with ggplot2 to a BentoBox layout}
\usage{
bb_plotGG(plot, x, y, width, height, just = c("left", "top"))
}
\arguments{
\item{plot}{ggplot object}

\item{x}{A unit object specifying x-location.}

\item{y}{A unit object specifying y-location.}

\item{width}{A unit object specifying width.}

\item{height}{A unit object specifying height.}

\item{just}{A string or numeric vector specifying the justification of the plot relative to its (x, y) location}
}
\description{
adds a plot created with ggplot2 to a BentoBox layout
}
