% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotManhattan.R
\name{bb_plotManhattan}
\alias{bb_plotManhattan}
\title{plots a Manhattan plot}
\usage{
bb_plotManhattan(
  bedfile,
  chrom,
  chromstart,
  chromend,
  col,
  ymax = 1,
  ylim = NULL,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = "center",
  default.units = "inches",
  draw = TRUE,
  ...
)
}
\arguments{
\item{bedfile}{bedfile for Manhattan plot}

\item{chrom}{chromosome of region to be plotted}

\item{chromstart}{chromstart of region to be plotted}

\item{chromend}{chromend of region to be plotted}

\item{col}{single colors, vector of colors, or color palette for coloring points}

\item{ymax}{fraction of max y value to set as height of plot}

\item{ylim}{numeric vector of length 2, giving the y coordinates range}

\item{x}{A numeric or unit object specifying x-location}

\item{y}{A numeric or unit object speicifying y-location}

\item{width}{A numeric or unit object specifying width}

\item{height}{A numeric or unit object specifying height}

\item{just}{justification of plot}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numeric vectors}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\description{
plots a Manhattan plot
}
