% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotManhattan.R
\name{bb_plotManhattan}
\alias{bb_plotManhattan}
\title{Plot a Manhattan plot}
\usage{
bb_plotManhattan(
  data,
  pVals,
  sigVal = 5e-08,
  chrom = NULL,
  chromstart = NULL,
  chromend = NULL,
  assembly = "hg19",
  sigLine = FALSE,
  sigCol = NULL,
  fill = "black",
  pch = 19,
  cex = 0.25,
  ymax = 1,
  range = NULL,
  space = 0.01,
  bg = NA,
  baseline = FALSE,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  params = NULL,
  ...
)
}
\arguments{
\item{data}{Data to be plotted; as a character value specifying a BED file path, a dataframe in BED format, or a \link[GenomicRanges]{GRanges} object.}

\item{pVals}{Character value specifying the name of the \code{data} column of corresponding p-values (will be converted to -log(10) space).}

\item{chrom}{Chromosome of region to be plotted, as a string.}

\item{chromstart}{Integer start position on chromosome to be plotted.}

\item{chromend}{Integer end position on chromosome to be plotted.}

\item{assembly}{Default genome assembly as a string or a \link[BentoBox]{bb_assembly} object. Default value is \code{assembly = "hg19"}.}

\item{sigLine}{Logical value indicating whether to draw a line at the significance level indicated with \code{sigVal}. Default value is \code{sigLine = FALSE}.}

\item{sigCol}{Single character value specifying the color of significant data points.}

\item{fill}{Character value(s) as a single value, vector, or palette specifying fill colors of data points. Default value is \code{fill = "black"}.}

\item{pch}{A numeric specifying point symbols. Default value is \code{pch = 19}.}

\item{cex}{A numeric indiciating the amount by which points should be scaled relative to the default. Default value is \code{cex = 0.25}.}

\item{ymax}{A numeric specifying the fraction of the max y-value to set as the height of the plot. Default value is \code{ymax = 1}.}

\item{range}{A numeric vector of length 2 specifying the y-range of p-values to plot (c(min, max)).}

\item{space}{A numeric value indicating the space between each chromsome as a fraction of the width of the plot, if plotting multiple chromosomes. Default value is \code{space = 0.01}.}

\item{bg}{Character value indicating background color. Default value is \code{bg = NA}.}

\item{baseline}{Logical value indicating whether to include a baseline along the x-axis. Default value is \code{baseline = FALSE}.}

\item{x}{A numeric or unit object specifying Manhattan plot x-location.}

\item{y}{A numeric or unit object specifying Manhattan plot y-location.}

\item{width}{A numeric or unit object specifying Manhattan plot width.}

\item{height}{A numeric or unit object specifying Manhattan plot height.}

\item{just}{Justification of Manhattan plot relative to its (x, y) location. If there are two values, the first value specifies horizontal justification and the second value specifies vertical justification.
Possible string values are: \code{"left"}, \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"}, and \code{"top"}. Default value is \code{just = c("left", "top")}.}

\item{default.units}{A string indicating the default units to use if \code{x}, \code{y}, \code{width}, or \code{height} are only given as numerics. Default value is \code{default.units = "inches"}.}

\item{draw}{A logical value indicating whether graphics output should be produced. Default value is \code{draw = TRUE}.}

\item{params}{An optional \link[BentoBox]{bb_assembly} object containing relevant function parameters.}

\item{...}{Additional grid graphical parameters. See \link[grid]{gpar}.}
}
\value{
Returns a \code{bb_manhattan} object containing relevant genomic region, placement, and \link[grid]{grob} information.
}
\description{
Plot a Manhattan plot
}
\details{
This function can be used to quickly plot a Manhattan plot by ignoring plot placement parameters:
\preformatted{
bb_plotManhattan(data, pVals,
                 chrom = NULL,
                 chromstart = NULL, chromend = NULL)
}
A Manhattan plot can be placed on a BentoBox coordinate page by providing plot placement parameters:
\preformatted{
bb_plotManhattan(data, pVals,
                 chrom = NULL,
                 chromstart = NULL, chromend = NULL,
                 x, y, width, height, just = c("left", "top"),
                 default.units = "inches")
}
}
\examples{
## Load GWAS data
data("bb_gwasData")

## Plot Manhattan plot filling up entire graphic device
bb_plotManhattan(data = bb_gwasData, pVals = "pVal", chrom = "chr21",
                 chromstart = 28000000, chromend = 30300000, ymax = 1.1, cex = 0.20)

## Plot and place Manhattan plot on a BentoBox page
bb_pageCreate(width = 5, height = 2, default.units = "inches")
manhattanPlot <- bb_plotManhattan(data = bb_gwasData, pVals = "pVal", chrom = "chr21",
                                  chromstart = 28000000, chromend = 30300000,
                                  ymax = 1.1, cex = 0.20,
                                  x = 0.5, y = 1.75, width = 4, height = 1.5,
                                  just = c("left", "bottom"), default.units = "inches")

## Annotate genome label
bb_annoGenomeLabel(plot = manhattanPlot, x = 0.5, y = 1.8)

## Annotate y-axis
bb_annoYaxis(plot = manhattanPlot, at = c(2, 4, 6), gp = gpar(fontsize = 10))
bb_plotText(label = "-log10(pVal)", fontsize = 8, rot = 90, x = 0.1, y = 0.75)

## Hide page guides
bb_pageGuideHide()

}
