% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotManhattan.R
\name{bb_plotManhattan}
\alias{bb_plotManhattan}
\title{plots a Manhattan plot}
\usage{
bb_plotManhattan(
  bed,
  pVals,
  params = NULL,
  chrom = NULL,
  chromstart = NULL,
  chromend = NULL,
  assembly = "hg19",
  fillcolor = "black",
  pch = 19,
  space = 0.01,
  cex = 0.25,
  ymax = 1,
  range = NULL,
  sigVal = 5e-08,
  sigLine = F,
  sigCol = NULL,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = T,
  ...
)
}
\arguments{
\item{bed}{data for Manhattan plot, can be any data in BED format or a GRanges}

\item{pVals}{name of column in bedfile of corresponding p-values (will be converted to -log(10) space)}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{chrom}{string of specific chromosome to zoom in on according to genome build (Ex: hg19 = "chr1"); if NULL all chromosomes found in data will be plotted}

\item{chromstart}{if single chromosome specified, start of region to be plotted}

\item{chromend}{if single chromosome specified, end of region to be plotted}

\item{assembly}{default genome assembly as a string or a bb_assembly object}

\item{space}{the space between each chromsome as a fraction of the width of the plot}

\item{cex}{number indiciating the amount by which points should be scaled relative to the default}

\item{ymax}{fraction of max y value to set as height of plot}

\item{range}{y-range of p-values to plot (c(min, max))}

\item{sigVal}{numeric indicating the significance level}

\item{sigLine}{logical indicating whether to draw a line at the significance level}

\item{sigCol}{single color for coloring significant values}

\item{x}{A numeric or unit object specifying x-location}

\item{y}{A numeric or unit object specifying y-location}

\item{width}{A numeric or unit object specifying width}

\item{height}{A numeric or unit object specifying height}

\item{just}{A string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numerics}

\item{draw}{A logical value indicating whether graphics output should be produced}

\item{colors}{single color, vector of colors, or color palette}

\item{transp}{transparency of colors}
}
\description{
plots a Manhattan plot
}
