% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotRaster.R
\name{bb_plotRaster}
\alias{bb_plotRaster}
\title{Plot a raster object within a BentoBox layout}
\usage{
bb_plotRaster(
  image,
  x,
  y,
  width,
  height,
  just = "center",
  default.units = "inches",
  interpolate = TRUE,
  params = NULL,
  ...
)
}
\arguments{
\item{image}{Any R object that can be coerced to a raster object.}

\item{x}{A numeric vector or unit object specifying raster x-locations.}

\item{y}{A numeric vector, unit object, or a character vector of values containing a "b" combined with a numeric value specifying raster y-locations.
The character vector will place raster y relative to the bottom of the most recently plotted BentoBox plot according to the units of the BentoBox page.}

\item{width}{A numeric vector or unit object specifying raster widths.}

\item{height}{A numeric vector or unit object specifying raster heights.}

\item{just}{Justification of text relative to its (x, y) location. If there are two values, the first value specifies horizontal justification and the second value specifies vertical justification.
Possible string values are: \code{"left"}, \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"}, and \code{"top"}. Default value is \code{just = "center"}.}

\item{default.units}{A string indicating the default units to use if \code{x}, \code{y}, \code{width}, or \code{height} are only given as numerics or numeric vectors.
Default value is \code{default.units = "inches"}.}

\item{interpolate}{A logical value indicating whether to linearly interpolate the image. Default value is \code{interpolate = TRUE}.}

\item{params}{An optional \link[BentoBox]{bb_params} object containing relevant function parameters.}

\item{...}{Additional grid graphical parameters. See \link[grid]{gpar}.}
}
\value{
Returns a \code{bb_raster} object containing relevant placement and \link[grid]{grob} information.
}
\description{
Plot a raster object within a BentoBox layout
}
\examples{

}
\seealso{
\link[grid]{grid.raster}
}
