% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotSignal.R
\name{bb_plotSignal}
\alias{bb_plotSignal}
\title{plots signal track data}
\usage{
bb_plotSignal(
  signal,
  chrom,
  params = NULL,
  chromstart = NULL,
  chromend = NULL,
  range = NULL,
  linecolor = "grey",
  binSize = NA,
  binCap = TRUE,
  fillcolor = NULL,
  assembly = "hg19",
  ymax = 1,
  scale = FALSE,
  bg = NA,
  baseline = FALSE,
  width = NULL,
  height = NULL,
  x = NULL,
  y = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  ...
)
}
\arguments{
\item{signal}{signal track data to be plotted (bigwig file, bed format dataframe, or GRanges with "counts" metadata); can be one or a list of two}

\item{chrom}{chromosome of region to be plotted as a string (i.e. "chr3")}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{chromstart}{start position}

\item{chromend}{end position}

\item{range}{y-range to plot (c(min, max))}

\item{linecolor}{color(s) of line outlining signal track(s); if using fillcolor and no outline is desired, set to NA}

\item{binSize}{the length of each bin in bp}

\item{binCap}{TRUE/FALSE whether the function will limit the number of bins to 8,000}

\item{fillcolor}{color(s) filling in signal track(s)}

\item{assembly}{desired genome assembly}

\item{ymax}{fraction of max y-value to set as height of plot}

\item{scale}{A logical value indicating whether to include a data scale label in the top left corner}

\item{bg}{background color}

\item{baseline}{logical indicating whether to include baseline along the x-axis}

\item{width}{A numeric or unit object specifying width}

\item{height}{A numeric or unit object specifying height}

\item{x}{A numeric or unit object specifying x-location}

\item{y}{A numeric or unit object specifying y-location}

\item{just}{a string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{default.units}{A string indicating the default units to use if x, y, width, or height are only given as numerics}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\description{
plots signal track data
}
