% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotSignal.R
\name{bb_plotSignal}
\alias{bb_plotSignal}
\title{plots signal track data}
\usage{
bb_plotSignal(signal, chrom, chromstart, chromend, range = NULL,
  linecolor = "grey", lwd = 1, yaxis = FALSE, binSize = NA,
  binCap = TRUE, fillcolor = NULL, transparency = NULL, ymax = 1,
  width = NULL, height = NULL, x = NULL, y = NULL,
  just = c("left", "top"), draw = TRUE, ...)
}
\arguments{
\item{signal}{signal track data to be plotted (bigwig file, bedgraph format dataframe, or bedgraph)}

\item{chrom}{chromosome of region to be plotted as a string (i.e. "chr3")}

\item{chromstart}{start position}

\item{chromend}{end position}

\item{range}{y-range to plot (c(min, max))}

\item{linecolor}{color of line outlining signal track}

\item{lwd}{linewidth of line outlining signal track}

\item{binSize}{the length of each bin in bp}

\item{binCap}{TRUE/FALSE whether the function will limit the number of bins to 8,000}

\item{fillcolor}{if want plot to be filled, specify fillcolor}

\item{transparency}{if want to specify a transparency for the fill, specify transparency}

\item{ymax}{fraction of max y value to set as height of plot}

\item{width}{A unit object specifying width}

\item{height}{A unit object specifying height}

\item{x}{A unit object specifying x-location}

\item{y}{A unit object specifying y-location}

\item{just}{a string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\description{
plots signal track data
}
