% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotTriangleHic.R
\name{bb_plotTriangleHic}
\alias{bb_plotTriangleHic}
\title{plots HiC interaction matrix in triangular format}
\usage{
bb_plotTriangleHic(
  hic,
  chrom = 8,
  chromstart = 133600000,
  chromend = 134800000,
  resolution = 10000,
  zrange = NULL,
  palette = colorRampPalette(c("white", "dark red")),
  width,
  x,
  y,
  just,
  norm = "KR",
  draw = T,
  ...
)
}
\arguments{
\item{hic}{path to .hic file or 3 column dataframe of counts}

\item{chrom}{chromosome of region to be plotted}

\item{chromstart}{chromosome start of region to be plotted}

\item{chromend}{chromosome end of region to be plotted}

\item{resolution}{the width in bp of each pixel; options are 2500000, 1000000, 500000, 250000, 100000, 50000, 25000, 10000, or 5000}

\item{zrange}{the range of interaction scores to plot, where extreme values will be set to the max or min}

\item{palette}{ColorRamp palette to use for representing interaction scores}

\item{width}{A unit object specifying the bottom width of the triangle}

\item{x}{A unit object specifying x-location}

\item{y}{A unit object specifying y-location}

\item{just}{a string or numeric vector specifying the justification of the viewport relative to its (x, y) location}

\item{norm}{if giving .hic file, hic data normalization; options are "NONE", "VC", "VC_SQRT", and "KR"}

\item{draw}{A logical value indicating whether graphics output should be produced}
}
\description{
plots HiC interaction matrix in triangular format
}
