% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_readBigwig.R
\name{bb_readBigwig}
\alias{bb_readBigwig}
\title{Reads a bigwig file and returns it as a data frame.}
\usage{
bb_readBigwig(
  filename,
  params = NULL,
  chrom = NULL,
  chromstart = 1,
  chromend = .Machine$integer.max,
  strand = "*"
)
}
\arguments{
\item{filename}{bigwig filename}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{chrom}{chromosome as a string, e.g. "chr3"}

\item{chromstart}{first base number to consider}

\item{chromend}{last base number to consider}

\item{strand}{strand, i.e. '+', '-', or '*'}
}
\value{
Function will return a 6-column dataframe of bigwig information.
}
\description{
Reads a bigwig file and returns it as a data frame.
}
