% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_readBigwig.R
\name{bb_readBigwig}
\alias{bb_readBigwig}
\title{Reads a bigwig file and returns it as a data frame.}
\usage{
bb_readBigwig(
  filename,
  chromosome = NULL,
  start = 1,
  end = .Machine$integer.max,
  strand = "*"
)
}
\arguments{
\item{filename}{bigwig filename}

\item{chromosome}{chromosome as a string, e.g. "chr3"}

\item{start}{first base number to consider}

\item{end}{last base number to consider}

\item{strand}{strand, i.e. '+', '-', or '*'}
}
\description{
Reads a bigwig file and returns it as a data frame.
}
