% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_readBigwig.R
\name{bb_readBigwig}
\alias{bb_readBigwig}
\title{Read a bigWig file and return it as a data frame}
\usage{
bb_readBigwig(
  file,
  chrom = NULL,
  chromstart = 1,
  chromend = .Machine$integer.max,
  strand = "*",
  params = NULL
)
}
\arguments{
\item{file}{A character value specifying the path to the bigwig file.}

\item{chrom}{Chromosome of data as a string, if data for a specific
chromosome is desired.}

\item{chromstart}{Integer start position on chromosome.}

\item{chromend}{Integer end position on chromosome.}

\item{strand}{A character value specifying strand.
Default value is \code{strand = "*"}. Options are:
\itemize{
\item{\code{"+"}: }{Plus strand.}
\item{\code{"-"}: }{Minus strand.}
\item{\code{"*"}: }{Plus and minus strands.}
}}

\item{params}{An optional \link[BentoBox]{bb_params} object
containing relevant function parameters.}
}
\value{
Returns a 6-column dataframe of bigwig information.
}
\description{
Read a bigWig file and return it as a data frame
}
