% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_readHic.R
\name{bb_readHic}
\alias{bb_readHic}
\title{extracts HiC data from .hic file using Straw}
\usage{
bb_readHic(
  hic,
  chrom,
  params = NULL,
  chromstart = NULL,
  chromend = NULL,
  resolution = "auto",
  zrange = NULL,
  norm = "KR",
  res_scale = "BP",
  assembly = "hg19",
  altchrom = NULL,
  altchromstart = NULL,
  altchromend = NULL
)
}
\arguments{
\item{hic}{path to .hic file}

\item{chrom}{chromosome of desired region}

\item{params}{an optional "bb_params" object space containing relevant function parameters}

\item{chromstart}{start position}

\item{chromend}{end position}

\item{resolution}{the width of each pixel; "auto" will attempt to choose a resolution (in basepairs) based on the size of the region}

\item{zrange}{the range of interaction scores to plot, where extreme values will be set to the max or min; if null, zrange will be set to (0, max(data))}

\item{norm}{hic data normalization; must be found in hic file}

\item{res_scale}{resolution scale; options are "BP" and "FRAG"}

\item{assembly}{desired genome assembly}

\item{altchrom}{if looking at region between two different chromosomes, this is the specified alternative chromosome}

\item{altchromstart}{if looking at region between two different chromosomes, start position of altchrom}

\item{altchromend}{if looking at region between two different chromsomes, end position of altchrom}
}
\value{
Function will return a 3-column dataframe
}
\description{
extracts HiC data from .hic file using Straw
}
