% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_removePlot.R
\name{bb_removePlot}
\alias{bb_removePlot}
\title{removes BentoBox plots}
\usage{
bb_removePlot(plot)
}
\arguments{
\item{plot}{BentoBox plot to remove; defined by the output of a BentoBox plotting function}
}
\description{
removes BentoBox plots
}
