% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_vGuide.R
\name{bb_vGuide}
\alias{bb_vGuide}
\title{draws a vertical guideline at a specified x-coordinate}
\usage{
bb_vGuide(x, units = "inches", col = "grey55")
}
\arguments{
\item{x}{x-coordinate to draw guide}

\item{units}{units of x-coordinate}

\item{col}{color of guideline}
}
\description{
draws a vertical guideline at a specified x-coordinate
}
