% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_vguide.R
\name{bb_vGuide}
\alias{bb_vGuide}
\title{draws a vertical guideline at a specified x-coordinate}
\usage{
bb_vGuide(x, col = "grey55", default.units = "inches")
}
\arguments{
\item{x}{A numeric or unit object specifying x-coordinate of guide}

\item{col}{color of guideline}

\item{default.units}{A string indicating the default units to use if x is only given as numeric vectors}
}
\description{
draws a vertical guideline at a specified x-coordinate
}
