% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCool.R
\name{.checkCool}
\alias{.checkCool}
\title{Check for .(m)cool file and contents}
\usage{
.checkCool(file)
}
\arguments{
\item{file}{Path to .(m)cool file}
}
\description{
Check for .(m)cool file and contents
}
\author{
Sarah Parker
}

\value{
A character string, either \code{".cool"} or \code{".mcool"}, indicating the file type. Aborts with an error if the file is not a valid .(m)cool file.
}
