% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCool.R
\name{.checkCoolErrors}
\alias{.checkCoolErrors}
\title{Error checking function for .(m)cool files}
\usage{
.checkCoolErrors(
  file,
  chrom,
  chromstart,
  chromend,
  zrange,
  altchrom,
  altchromstart,
  altchromend,
  norm,
  resolution
)
}
\description{
Error checking function for .(m)cool files
}
