% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCool.R
\name{.coolAutoResolution}
\alias{.coolAutoResolution}
\title{Determine best resolution for size of region for .(m)cool files}
\usage{
.coolAutoResolution(file, chromstart, chromend)
}
\description{
Determine best resolution for size of region for .(m)cool files
}
