% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCool.R
\name{.coolRegion}
\alias{.coolRegion}
\title{Add (alt)chromstart and (alt)chromend for NULL (alt)chrom region of 
.(m)cool files}
\usage{
.coolRegion(file, chrom, resolution)
}
\description{
Add (alt)chromstart and (alt)chromend for NULL (alt)chrom region of 
.(m)cool files
}
