% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCool.R
\name{.pullBinChunks}
\alias{.pullBinChunks}
\title{Read in data for a bin chunk}
\usage{
.pullBinChunks(
  binChunk,
  file,
  bin_offsets,
  binChunkSize,
  datasetPath,
  end1bin,
  start2bin,
  end2bin
)
}
\arguments{
\item{binChunk}{The binChunk indeces to read}

\item{file}{Path to .(m)cool file}

\item{bin_offsets}{Read in bin1 offsets}

\item{binChunkSize}{Size of bin chunk, for comparison against the end of 
the bin chunk}

\item{datasetPath}{Dataset path, for specifying resolution in .mcool file}

\item{end1bin}{Bin where end1 starts}

\item{start2bin}{Bin for chr2 starts}

\item{end2bin}{Bin for end2 starts}
}
\description{
Read in data for a bin chunk
}
