% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHicTriangle.R
\name{plotHicTriangle}
\alias{plotHicTriangle}
\title{Plot a Hi-C interaction matrix in a triangular format}
\usage{
plotHicTriangle(
    data,
    resolution = "auto",
    zrange = NULL,
    norm = "KR",
    matrix = "observed",
    chrom,
    chromstart = NULL,
    chromend = NULL,
    assembly = "hg38",
    palette = colorRampPalette(brewer.pal(n = 9, "YlGnBu")),
    colorTrans = "linear",
    flip = FALSE,
    bg = NA,
    x = NULL,
    y = NULL,
    width = NULL,
    height = NULL,
    just = c("left", "top"),
    default.units = "inches",
    draw = TRUE,
    params = NULL,
    quiet = FALSE
)
}
\arguments{
\item{data}{A string path to .hic file as a string or a 3-column
dataframe of interaction counts in sparse upper triangular format. Class: Data.}

\item{resolution}{A numeric specifying the width in basepairs of
each pixel. For hic files, "auto" will attempt to choose a
resolution based on the size of the region. For
dataframes, "auto" will attempt to detect the resolution
the dataframe contains. Default value is \code{resolution = "auto"}. Class: Data.}

\item{zrange}{A numeric vector of length 2 specifying the
range of interaction scores to plot, where extreme values
will be set to the max or min. Default value is \code{zrange = NULL}. Class: Data.}

\item{norm}{A character value specifying hic data normalization method,
if giving .hic file. This value must be found in the .hic file.
Default value is \code{norm = "KR"}. Class: Data.}

\item{matrix}{A character value indicating the type of matrix to output.
Default value is \code{matrix = "observed"}. Options include: "observed", "oe", "log2oe". Class: Data.
\itemize{
\item{\code{"observed"}: }{Observed counts.}
\item{\code{"oe"}: }{Observed/expected counts.}
\item{\code{"log2oe"}: }{Log2 transformed observed/expected counts.}
}}

\item{chrom}{A string representing the chromosome of region to be plotted. Class: Data.}

\item{chromstart}{A integer start position on chromosome to be plotted. Default value is \code{chromstart = NULL}. Class: Data.}

\item{chromend}{A integer end position on chromosome to be plotted. Default value is \code{chromend = NULL}. Class: Data.}

\item{assembly}{A string representing the eefault genome assembly or a
\link[plotgardener]{assembly} object.
Default value is \code{assembly = "hg38"}. Class: Data.}

\item{palette}{A function describing the color palette to use for
representing scale of interaction scores. Default value is
\code{palette =  colorRampPalette(brewer.pal(n = 9, "YlGnBu"))}. Class: Aesthetic.}

\item{colorTrans}{A string specifying how to scale Hi-C colors.
Default value is \code{colorTrans = "linear"}. Options include "linear", "log", "log2", or "log10". Class: Aesthetic.}

\item{flip}{A logical indicating whether to flip the orientation of
the Hi-C matrix over the x-axis. Default value is \code{flip = FALSE}. Options include: TRUE, FALSE. Class: Aesthetic.}

\item{bg}{A character value indicating background color.
Default value is \code{bg = NA}. Class: Aesthetic.}

\item{x}{A numeric or unit object specifying triangle Hi-C plot x-location. Class: Positional.}

\item{y}{A numeric, unit object, or character containing a "b"
combined with a numeric value specifying triangle Hi-C plot y-location.
The character value will
place the triangle Hi-C plot y relative to the bottom of the most
recently plotted plot according to the units of the plotgardener page. Class: Positional.}

\item{width}{A numeric or unit object specifying the bottom
width of the Hi-C plot triangle. Class: Positional.}

\item{height}{A numeric or unit object specifying the height of
the Hi-C plot triangle. Class: Positional.}

\item{just}{A string specifying the justification of triangle Hi-C plot relative to
its (x, y) location. If there are two values, the first value specifies
horizontal justification and the second value specifies vertical
justification.
Possible string values are: \code{"left"}, \code{"right"},
\code{"centre"}, \code{"center"}, \code{"bottom"}, and \code{"top"}.
Default value is \code{just = c("left", "top")}. Options include: c("left", "top"), c("left", "bottom"), "left", c("right", "top"), c("right", "bottom"), "right", "top", "bottom", "center". Class: Positional.}

\item{default.units}{A string indicating the default units to use if
\code{x}, \code{y}, \code{width}, or \code{height} are only given as
numerics. Default value is \code{default.units = "inches"}. Options include: "inches", "cm", "npc", "snpc", "native", "mm", "points". Class: Positional.}

\item{draw}{A logical value indicating whether graphics output should
be produced. Default value is \code{draw = TRUE}. Options include: TRUE, FALSE. Class: Positional.}

\item{params}{An optional \link[plotgardener]{pgParams} object containing
relevant function parameters. Class: Positional.}

\item{quiet}{A logical indicating whether or not to print messages. Default value is \code{quiet = FALSE}. Options include: TRUE, FALSE. Class: Positional.

\end}
}
\value{
Returns a \code{hicTriangle} object containing relevant
genomic region, Hi-C data, placement, and \link[grid]{grob} information.
}
\description{
Plot a Hi-C interaction matrix in a triangular format
}
\details{
In this orientation, the x-axis represents the genomic coordinates and
the y-axis corresponds to distance in Hi-C bins.

A triangle Hi-C plot can be placed on a plotgardener coordinate page
by providing plot placement parameters:
\preformatted{
plotHicTriangle(data, chrom,
                chromstart = NULL, chromend = NULL,
                x, y, width, height, just = c("left", "top"),
                default.units = "inches")
}
This function can also be used to quickly plot an unannotated triangle
Hi-C plot by ignoring plot placement parameters:
\preformatted{
plotHicTriangle(data, chrom,
                chromstart = NULL, chromend = NULL)
}

If \code{height} is \eqn{<} \eqn{0.5 * width}, the top of the triangle
will be cropped to the given \code{height}.
}
\examples{
## Load Hi-C data
library(plotgardenerData)
data("IMR90_HiC_10kb")

## Create a page
pageCreate(width = 4, height = 2.5, default.units = "inches")

## Plot and place triangle Hi-C plot
hicPlot <- plotHicTriangle(
    data = IMR90_HiC_10kb, resolution = 10000,
    zrange = c(0, 70),
    chrom = "chr21",
    chromstart = 28000000, chromend = 30300000,
    assembly = "hg19",
    x = 2, y = 0.5, width = 3, height = 1.5,
    just = "top", default.units = "inches"
)

## Annotate x-axis genome label
annoGenomeLabel(
    plot = hicPlot, scale = "Mb", x = 0.5, y = 2.03,
    just = c("left", "top")
)

## Annotate heatmap legend
annoHeatmapLegend(
    plot = hicPlot, x = 3.5, y = 0.5,
    width = 0.13, height = 1.2,
    just = c("right", "top")
)

## Hide page guides
pageGuideHide()
}
\seealso{
\link[plotgardener]{readHic}
}
