% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiSignal.R
\name{plotMultiSignal}
\alias{plotMultiSignal}
\title{Plot multiple signal tracks inline with each other}
\usage{
plotMultiSignal(
   data, 
   x = NULL, 
   y = NULL,
   orientation = "h", 
   width = NULL,
   height = NULL, 
   gapdistance = .2,
   just = c("left", "top"),
   default.units = "inches", 
   binSize = NA, 
   binCap = TRUE, 
   negData = FALSE,
   chrom, 
   chromstart = NULL, 
   chromend = NULL,
   assembly = "hg38", 
   linecolor,
   fill = NA, 
   ymax = 1, 
   range = NULL, 
   scale = FALSE,
   bg = NA, 
   baseline = TRUE, 
   baseline.color = "grey",
   baseline.lwd = 1, 
   draw = TRUE,
   params = NULL, ...
)
}
\arguments{
\item{data}{List of data to be plotted as a character value specifying a
bigwig file path, a dataframe in BED format, or a
\link[GenomicRanges]{GRanges} object with metadata column \code{score}.
Either one \code{data} argument or a list of two can be provided, where
the second \code{data} will be plotted below the x-axis.}

\item{binCap}{A logical value indicating whether the function will
limit the number of data bins to 8,000.
Default value is \code{binCap = TRUE}.}

\item{negData}{A logical value indicating whether the data has both
positive and negative scores and the y-axis should be split.
Default value is \code{negData = FALSE}.}

\item{chrom}{Chromosome of region to be plotted, as a string.}

\item{chromstart}{Integer start position on chromosome to be plotted.}

\item{chromend}{Integer end position on chromosome to be plotted.}

\item{assembly}{Default genome assembly as a string or a
\link[plotgardener]{assembly} object.
Default value is \code{assembly = "hg38"}.}

\item{linecolor}{A character value or vector of character values specifying the
line color(s) outlining the signal track(s).
Default value is \code{linecolor = "#37a7db"}.}

\item{fill}{A character value or vector of length 2 specifying
the fill color(s) of the signal track(s). Default value is \code{fill = NA}.}

\item{ymax}{A numeric specifying the fraction of the max y-value
to set as the height of the plot. Default value is \code{ymax = 1}.}

\item{range}{A numeric vector of length 2 specifying the y-range
of data to plot (c(min, max)).}

\item{scale}{A logical value indicating whether to include a data
scale label in the top left corner of the plot.
Default value is \code{scale = FALSE}.}

\item{bg}{Character value indicating background color.
Default value is \code{bg = NA}.}

\item{baseline}{Logical value indicating whether to include a
baseline along the x-axis. Default value is \code{baseline = TRUE}.}

\item{baseline.color}{Baseline color.
Default value is \code{baseline.color = "grey"}.}

\item{baseline.lwd}{Baseline line width.
Default value is \code{baseline.lwd = 1}.}

\item{orientation}{A string specifying signal track orientation.
Default value is \code{orientation = "h"}. Options are:
\itemize{
\item{\code{"v"}: }{Vertical signal track orientation.}
\item{\code{"h"}: }{Horizontal signal track orientation.}
}}

\item{x}{A numeric vector or unit object specifying the starting
vertex x-locations.}

\item{y}{A numeric vector, unit object, or a character vector
of values containing a "b" combined with a numeric value specifying
polygon vertex y-locations.
The character vector will place polygon vertex y-locations relative
to the bottom of the most recently plotted plot according
to the units of the plotgardener page.}

\item{width}{A numeric or unit object specifying signal plot width.}

\item{height}{A numeric or unit object specifying signal plot height.}

\item{just}{Justification of signal plot relative to its (x, y) location.
If there are two values, the first value specifies horizontal justification
and the second value specifies vertical justification.
Possible string values are: \code{"left"}, \code{"right"},
\code{"centre"}, \code{"center"}, \code{"bottom"}, and \code{"top"}.
Default value is \code{just = c("left", "top")}.}

\item{default.units}{A string indicating the default units to use
if \code{x} or \code{y} are only given as numeric vectors.
Default value is \code{default.units = "inches"}.
#' @param binSize A numeric specifying the length of each data
bin in basepairs. Default value is \code{binSize = NA}.}

\item{gapdistance}{A numeric object specifying space between plots}

\item{draw}{A logical value indicating whether graphics output should be
produced. Default value \code{draw = TRUE}.}

\item{params}{An optional \link[plotgardener]{pgParams} object containing
relevant function parameters.}

\item{...}{Additional grid graphical parameters. See \link[grid]{gpar}.}
}
\value{
Returns several \code{polygon} object containing relevant
placement and \link[grid]{grob} information.
}
\description{
Plot multiple signal tracks inline with each other
}
\examples{
library("plotgardener")
library("org.Hs.eg.db")
library("TxDb.Hsapiens.UCSC.hg19.knownGene")
library("plotgardenerData")
data("GM12878_HiC_10kb")
data("IMR90_HiC_10kb")
data("GM12878_ChIP_CTCF_signal")
data("IMR90_ChIP_CTCF_signal")
data("GM12878_ChIP_H3K27ac_signal")
data("IMR90_ChIP_H3K27ac_signal")


## Load libraries and datasets
testList <- list(GM12878_ChIP_CTCF_signal, GM12878_ChIP_H3K27ac_signal, IMR90_ChIP_CTCF_signal, IMR90_ChIP_H3K27ac_signal)
## Set genomic and dimension parameters in a `params` object
params <- pgParams(chrom = "chr21", chromstart = 28150000, chromend = 29150000, 
                    assembly = "hg19", x = 3.5, width = 1.5, default.units = "inches")

pageCreate(width = 7, height = 6, default.units = "inches")

plotMultiSignal(testList, chrom = "chr21", assembly = "hg19", params = c(params_c, ctcf_range),
               fill = "#253494",y = 0.2, height = 4,  x = 0.2, width = 2, gapdistance = 0.1, orientation = "h")
}
