% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCool.R
\name{readCool}
\alias{readCool}
\title{Read a .(m)cool file and return Hi-C data as a dataframe}
\usage{
readCool(
    file,
    chrom,
    chromstart = NULL,
    chromend = NULL,
    altchrom = NULL,
    altchromstart = NULL,
    altchromend = NULL,
    assembly = "hg38",
    resolution = "auto",
    res_scale = "BP",
    zrange = NULL,
    norm = "NONE",
    matrix = "observed",
    params = NULL,
    quiet = FALSE
)
}
\description{
Read a .(m)cool file and return Hi-C data as a dataframe
}
\author{
Sarah Parker, Nicole Kramer
}
